Function to get core dataset.

This function executes either the s3_select query or data download and local merge on a specified subset of the core dataset. It then merges the dataset with the ntee dataframe.

Usage

get_core(dsname, time, scope.orgtype, scope.formtype, filters, append.bmf)

Arguments

dsname

character scalar. Name of data series to query from S3. Valid inputs are "core" and "bmf", not both.

time

character vector. Dates of core/bmf files to query. Valid inputs range from 1989-2022. Default value is "current" for 2022.

scope.orgtype

character scalar. Organization type to query from core/bmf s3 bucket. Valid inputs are 'CHARITIES' for charities (501C3-PC), 'PRIVFOUND' for private foundations (501C3-PF) and 'NONPROFIT' for all nonprofits (501CE)

scope.formtype

character scalar. Form type to query from core/bmf s3 bucket. Valid inputs are 'PC'(nonprofits that file the full version), 'EZ'(nonprofits that file 990EZs only), ' PZ'(nonprofits that file both PC and EZ), or 'PF'(private foundations).

filters

list. List of column filters to apply

append.bmf

boolean. Option to merge queried core data with bmf data. Involves downloading the bmf dataset and will take longer. Default == FALSE.

Value

a fully merged core data.table for the end user