Script with front end functions for users to access NTEE2 codes and filter NTEE dataset with codes

This function takes user inputs across levels 1- 5 of the NTEE2 code structure and returns selected columns from the NTEE database. The use can choose to return a dataframe or reactable object.

Usage

ntee_preview(ntee.group, ntee.code, ntee.orgtype, cols, visualize)

Arguments

ntee.group

character vector. Vector of desired Industry Group codes to filter. Use "all" to include all possible codes.

ntee.code

character vector. Sequence of desired Industry, Division and Subdivision codes (old code structure) to use in filtering. Use "all" to include all possible codes. Can also provide only partial codes. For example "A" or "Axx" will query NTEE2 codes based on Industry group "A" and all division and subdivisions.

ntee.orgtype

character vector. Vector of Organization Types. Use "all" to include all possible codes.

cols

character vector. A vector containing list of columns to select. Default value is "all" for all columns.

visualize

boolean. A boolean value indicating whether to return a reactable object or data frame.

Value

a dataframe or reactable object with rows that match user arguments.

Details

Function to interactively explore NTEE codes

Examples

ntee_preview("all", "A1x", "all", "all",TRUE)
#> Collecting Matching Industry Groups
#> Collecting Matching Industry Division and Subdivisions
#> Collecting Matching Organization Types

ntee_preview("ART", "Axx", "all", "all",TRUE)
#> Collecting Matching Industry Groups
#> Collecting Matching Industry Division and Subdivisions
#> Collecting Matching Organization Types

ntee_preview("all", "Axx", "all", c("type.org", "univ"), FALSE)
#> Collecting Matching Industry Groups
#> Collecting Matching Industry Division and Subdivisions
#> Collecting Matching Organization Types
#>    type.org  univ
#> 1        AA FALSE
#> 2        AA FALSE
#> 3        MT FALSE
#> 4        MT FALSE
#> 5        MT FALSE
#> 6        MT FALSE
#> 7        MT FALSE
#> 8        MT FALSE
#> 9        PA FALSE
#> 10       PA FALSE
#> 11       PA FALSE
#> 12       PA FALSE
#> 13       PA FALSE
#> 14       PA FALSE
#> 15       PA FALSE
#> 16       PA FALSE
#> 17       PA FALSE
#> 18       PA FALSE
#> 19       PA FALSE
#> 20       PA FALSE
#> 21       PA FALSE
#> 22       PA FALSE
#> 23       RP FALSE
#> 24       RP FALSE
#> 25       RP FALSE
#> 26       MS FALSE
#> 27       MS FALSE
#> 28       MS FALSE
#> 29       MS FALSE
#> 30       MS FALSE
#> 31       MS FALSE
#> 32       MS FALSE
#> 33       MS FALSE
#> 34       MS FALSE
#> 35       MS FALSE
#> 36       MS FALSE
#> 37       MS FALSE
#> 38       MS FALSE
#> 39       MS FALSE
#> 40       MS FALSE
#> 41       MM FALSE
#> 42       MM FALSE
#> 43       MM FALSE
#> 44       MM FALSE
#> 45       MM FALSE
#> 46       MM FALSE
#> 47       MM FALSE
#> 48       NS FALSE
#> 49       NS FALSE
#> 50       NS FALSE
#> 51       NS FALSE
#> 52       RG FALSE
#> 53       RG FALSE
#> 54       RG FALSE
#> 55       RG FALSE
#> 56       RG FALSE
#> 57       RG FALSE
#> 58       RG FALSE
#> 59       RG FALSE
#> 60       RG FALSE
#> 61       RG FALSE
#> 62       RG FALSE
#> 63       RG FALSE
#> 64       RG FALSE
#> 65       RG FALSE
#> 66       RG FALSE
#> 67       RG FALSE
#> 68       RG FALSE
#> 69       RG FALSE
#> 70       RG FALSE
#> 71       RG FALSE
#> 72       RG FALSE
#> 73       RG FALSE
#> 74       RG FALSE
#> 75       RG FALSE
#> 76       RG FALSE
#> 77       RG FALSE
#> 78       RG FALSE
#> 79       RG FALSE
#> 80       RG FALSE
#> 81       RG FALSE
#> 82       RG FALSE
#> 83       RG FALSE
#> 84       RG FALSE
#> 85       RG FALSE
#> 86       RG FALSE
#> 87       RG FALSE