Function that returns FIPS codes that match dynamic User arguments

Filters either the Block or Tract data.tables to return a list of FIPS codes that match conditions specified by the User

Usage

parse_fips(census.level, ...)

Arguments

census.level

string. data.table to parse; "BLOCK" | "TRACT"

...

columns in either tract or block dataset for filtering.

Value

a list of FIPS codes for either Tract IDs or Block IDs.

Examples

 parse_fips(census.level = "TRACT",
                     state.census.abbr = c("NY", "MD"))
#> Warning: There was 1 warning in `dplyr::filter()`.
#> ℹ In argument: `state.census.abbr == c("NY", "MD")`.
#> Caused by warning in `state.census.abbr == c("NY", "MD")`:
#> ! longer object length is not a multiple of shorter object length
#>    [1] "24001000100" "24001000300" "24001000500" "24001000700" "24001001000"
#>    [6] "24001001200" "24001001401" "24001001502" "24001001600" "24001001800"
#>   [11] "24001002000" "24001002200" "24003702205" "24003702208" "24003740102"
#>   [16] "24003740104" "24003740201" "24003740303" "24003740305" "24003740500"
#>   [21] "24003740602" "24003740701" "24003740800" "24003741000" "24003751400"
#>   [26] "24003751700" "24003730100" "24003730204" "24003730401" "24003731308"
#>   [31] "24003731310" "24003750101" "24003750201" "24003750203" "24003750400"
#>   [36] "24003750803" "24003750900" "24003751102" "24003730502" "24003730505"
#>   [41] "24003730601" "24003730604" "24003730800" "24003730902" "24003731003"
#>   [46] "24003731102" "24003731104" "24003731201" "24003731203" "24003731303"
#>   [51] "24003731307" "24003701102" "24003701300" "24003702100" "24003702402"
#>   [56] "24003702601" "24003702701" "24003706101" "24003706302" "24003706402"
#>   [61] "24003706600" "24003707001" "24003708001" "24003751600" "24005402201"
#>   [66] "24005402503" "24005403701" "24005403802" "24005404800" "24005405000"
#>   [71] "24005407001" "24005408100" "24005408302" "24005408304" "24005410200"
#>   [76] "24005411102" "24005411202" "24005451000" "24005451801" "24005451803"
#>   [81] "24005402302" "24005402306" "24005402406" "24005402505" "24005402509"
#>   [86] "24005402603" "24005403702" "24005404102" "24005404202" "24005404403"
#>   [91] "24005404501" "24005492401" "24005401200" "24005402305" "24005402404"
#>   [96] "24005403100" "24005403202" "24005403401" "24005403500" "24005403602"
#>  [101] "24005408400" "24005408503" "24005408506" "24005408601" "24005408702"
#>  [106] "24005408704" "24005408900" "24005490400" "24005411404" "24005440100"
#>  [111] "24005440300" "24005490301" "24005490601" "24005490603" "24005490701"
#>  [116] "24005490800" "24005491000" "24005491201" "24005491300" "24005491402"
#>  [121] "24005491600" "24005491900" "24005492002" "24005492102" "24005492600"
#>  [126] "24005411306" "24005411308" "24005411406" "24005411409" "24005440400"
#>  [131] "24005440600" "24005440702" "24005440900" "24005441101" "24005450100"
#>  [136] "24005451300" "24005451402" "24005451701" "24005420200" "24005420302"
#>  [141] "24005420401" "24005420500" "24005420701" "24005420800" "24005421000"
#>  [146] "24005421102" "24005421300" "24005450300" "24005450501" "24005450504"
#>  [151] "24005450900" "24005451500" "24005452100" "24005452400" "24005492300"
#>  [156] "24005400100" "24005400400" "24005400600" "24005400702" "24005400900"
#>  [161] "24005401101" "24005401301" "24005401400" "24005401504" "24005401506"
#>  [166] "24005430101" "24005430200" "24005430400" "24005430700" "24005430900"
#>  [171] "24005980100" "24009860101" "24009860200" "24009860401" "24009860501"
#>  [176] "24009860600" "24009860702" "24009860801" "24009860900" "24009861003"
#>  [181] "24009990100" "24011955100" "24011955202" "24011955302" "24011955500"
#>  [186] "24013501001" "24013502000" "24013504100" "24013504202" "24013505102"
#>  [191] "24013505205" "24013505207" "24013506101" "24013506200" "24013507601"
#>  [196] "24013507702" "24013507704" "24013507802" "24013508102" "24013509001"
#>  [201] "24013510000" "24013512000" "24013513002" "24013514201" "24015030100"
#>  [206] "24015030400" "24015030503" "24015030506" "24015030602" "24015030903"
#>  [211] "24015030905" "24015031201" "24015031301" "24015031400" "24017850102"
#>  [216] "24017850202" "24017850400" "24017850600" "24017850708" "24017850710"
#>  [221] "24017850712" "24017850801" "24017850901" "24017850904" "24017850906"
#>  [226] "24017851002" "24017851200" "24017851302" "24017851500" "24019970100"
#>  [231] "24019970300" "24019970500" "24019970702" "24019970900" "24021740200"
#>  [236] "24021751600" "24021751801" "24021751903" "24021752101" "24021752201"
#>  [241] "24021752204" "24021752303" "24021752502" "24021752602" "24021752801"
#>  [246] "24021752900" "24021753002" "24021767500" "24021770700" "24021775302"
#>  [251] "24021750100" "24021750300" "24021750504" "24021750506" "24021750701"
#>  [256] "24021750801" "24021750803" "24021751002" "24021751004" "24021751202"
#>  [261] "24021751301" "24021751901" "24021751904" "24021765100" "24021775600"
#>  [266] "24023000200" "24023000400" "24023000600" "24025301102" "24025301202"
#>  [271] "24025302200" "24025303102" "24025303203" "24025303205" "24025303300"
#>  [276] "24025303501" "24025303602" "24025303605" "24025303700" "24025303802"
#>  [281] "24025303900" "24025304102" "24025304202" "24025305200" "24025301105"
#>  [286] "24025301107" "24025301204" "24025301301" "24025301401" "24025301601"
#>  [291] "24025301702" "24025301704" "24025302801" "24025302901" "24025306100"
#>  [296] "24025306300" "24025306500" "24027602201" "24027602303" "24027602305"
#>  [301] "24027603001" "24027603004" "24027604002" "24027605103" "24027605401"
#>  [306] "24027605502" "24027605504" "24027605601" "24027601103" "24027601105"
#>  [311] "24027601108" "24027601203" "24027602302" "24027602700" "24027602900"
#>  [316] "24027606603" "24027606606" "24027606701" "24027606705" "24027606707"
#>  [321] "24027606804" "24027606806" "24027606904" "24027606906" "24029950100"
#>  [326] "24029950300" "24029950500" "24031700101" "24031700204" "24031700206"
#>  [331] "24031700208" "24031700312" "24031700500" "24031700608" "24031700613"
#>  [336] "24031701212" "24031701306" "24031701308" "24031701313" "24031701315"
#>  [341] "24031701317" "24031700105" "24031700306" "24031700309" "24031700713"
#>  [346] "24031700716" "24031700722" "24031700811" "24031700813" "24031700818"
#>  [351] "24031700830" "24031700833" "24031700835" "24031700607" "24031700615"
#>  [356] "24031700704" "24031700710" "24031700717" "24031700719" "24031700723"
#>  [361] "24031700816" "24031700820" "24031700823" "24031700826" "24031700829"
#>  [366] "24031700902" "24031700904" "24031701001" "24031701004" "24031701006"
#>  [371] "24031701101" "24031701210" "24031701220" "24031701201" "24031701205"
#>  [376] "24031701213" "24031701215" "24031701218" "24031704401" "24031704404"
#>  [381] "24031704502" "24031704600" "24031704803" "24031704805" "24031705000"
#>  [386] "24031705200" "24031705400" "24031705502" "24031705602" "24031705702"
#>  [391] "24031705901" "24031705903" "24031706007" "24031706009" "24031706011"
#>  [396] "24031706013" "24031703201" "24031703206" "24031703209" "24031703213"
#>  [401] "24031703215" "24031703218" "24031703301" "24031703401" "24031703403"
#>  [406] "24031703501" "24031703601" "24031703701" "24031703800" "24031704100"
#>  [411] "24031704300" "24031701408" "24031701410" "24031701415" "24031701418"
#>  [416] "24031701421" "24031701423" "24031701505" "24031701507" "24031701509"
#>  [421] "24031703210" "24031703221" "24031701602" "24031701702" "24031701704"
#>  [426] "24031701900" "24031702101" "24031702200" "24031702302" "24031702402"
#>  [431] "24031702601" "24031702700" "24031702900" "24031703100" "24031704000"
#>  [436] "24033805101" "24033805202" "24033805601" "24033805700" "24033805802"
#>  [441] "24033805906" "24033805908" "24033806000" "24033806713" "24033806800"
#>  [446] "24033807000" "24033807301" "24033800102" "24033800105" "24033800108"
#>  [451] "24033800203" "24033800208" "24033800210" "24033800212" "24033800214"
#>  [456] "24033800410" "24033800412" "24033803606" "24033806710" "24033806712"
#>  [461] "24033807404" "24033807407" "24033807409" "24033803605" "24033803612"
#>  [466] "24033803700" "24033803803" "24033804001" "24033804101" "24033804200"
#>  [471] "24033804400" "24033804700" "24033804802" "24033806100" "24033806300"
#>  [476] "24033806601" "24033806706" "24033802002" "24033802104" "24033802107"
#>  [481] "24033802204" "24033802404" "24033802407" "24033802502" "24033802700"
#>  [486] "24033802804" "24033802901" "24033803002" "24033803200" "24033803401"
#>  [491] "24033803508" "24033803519" "24033803525" "24033800401" "24033800403"
#>  [496] "24033800409" "24033800505" "24033800509" "24033800513" "24033800515"
#>  [501] "24033800517" "24033800519" "24033800601" "24033800605" "24033800607"
#>  [506] "24033800701" "24033803512" "24033803514" "24033803520" "24033803522"
#>  [511] "24033803526" "24033803601" "24033803608" "24033800705" "24033800707"
#>  [516] "24033800900" "24033801004" "24033801006" "24033801207" "24033801209"
#>  [521] "24033801211" "24033801213" "24033801215" "24033801217" "24033801305"
#>  [526] "24033801308" "24033801310" "24033801312" "24033801404" "24033801406"
#>  [531] "24033801408" "24033801410" "24033801500" "24033801701" "24033801704"
#>  [536] "24033801707" "24033801801" "24033801804" "24033801807" "24033801901"
#>  [541] "24033801905" "24033801907" "24033802001" "24033802501" "24035810200"
#>  [546] "24035810400" "24035810600" "24035810800" "24035810902" "24035990000"
#>  [551] "24035990200" "24037875100" "24037875202" "24037875400" "24037875600"
#>  [556] "24037875801" "24037875901" "24037876001" "24037876100" "24037990000"
#>  [561] "24039930102" "24039930300" "24039930600" "24039990100" "24041960201"
#>  [566] "24041960400" "24041960502" "24041960700" "24041960900" "24043000100"
#>  [571] "24043000301" "24043000400" "24043000601" "24043000700" "24043000900"
#>  [576] "24043001002" "24043010200" "24043010400" "24043010600" "24043010801"
#>  [581] "24043010900" "24043011100" "24043011202" "24043011302" "24043011500"
#>  [586] "24045000100" "24045000300" "24045000500" "24045010102" "24045010300"
#>  [591] "24045010501" "24045010603" "24045010605" "24045010701" "24045010800"
#>  [596] "24047950100" "24047950400" "24047950700" "24047950900" "24047951100"
#>  [601] "24047951300" "24047951500" "24047980000" "24510130805" "24510150100"
#>  [606] "24510150300" "24510150500" "24510150701" "24510150800" "24510151000"
#>  [611] "24510151200" "24510160100" "24510160300" "24510160500" "24510160700"
#>  [616] "24510271600" "24510271801" "24510271900" "24510272004" "24510272006"
#>  [621] "24510280101" "24510280200" "24510280302" "24510090300" "24510090500"
#>  [626] "24510120100" "24510120202" "24510120400" "24510120700" "24510130200"
#>  [631] "24510130400" "24510130700" "24510130804" "24510140200" "24510170200"
#>  [636] "24510270804" "24510271001" "24510271101" "24510271200" "24510271400"
#>  [641] "24510080101" "24510080500" "24510090600" "24510260101" "24510260201"
#>  [646] "24510260301" "24510270101" "24510270200" "24510270302" "24510270402"
#>  [651] "24510270502" "24510270701" "24510270703" "24510270802" "24510270901"
#>  [656] "24510270903" "24510010200" "24510010400" "24510020100" "24510020300"
#>  [661] "24510030200" "24510040200" "24510060200" "24510060400" "24510070200"
#>  [666] "24510070400" "24510080301" "24510080400" "24510080700" "24510090900"
#>  [671] "24510100200" "24510110100" "24510120500" "24510220100" "24510260303"
#>  [676] "24510260402" "24510260404" "24510260604" "24510260700" "24510260900"
#>  [681] "24510261100" "24510160801" "24510180100" "24510180300" "24510190200"
#>  [686] "24510200100" "24510200300" "24510200500" "24510200701" "24510200800"
#>  [691] "24510210200" "24510230200" "24510240100" "24510240300" "24510250101"
#>  [696] "24510250103" "24510250204" "24510250206" "24510250301" "24510250401"
#>  [701] "24510250500" "24510280401" "24510280403" "36001000200" "36001000401"
#>  [706] "36001000404" "36001000502" "36001000700" "36001001100" "36001001500"
#>  [711] "36001001700" "36001001802" "36001001902" "36001002100" "36001002300"
#>  [716] "36001002600" "36001014002" "36001012800" "36001013000" "36001013200"
#>  [721] "36001013400" "36001013505" "36001013507" "36001013601" "36001013703"
#>  [726] "36001013706" "36001013801" "36001013901" "36001014100" "36001014202"
#>  [731] "36001014301" "36001014401" "36001014501" "36001014503" "36001014607"
#>  [736] "36001014609" "36001014612" "36001014614" "36001014700" "36001014802"
#>  [741] "36003940200" "36003950300" "36003950500" "36003950700" "36003950900"
#>  [746] "36003951100" "36003951300" "36005026702" "36005027700" "36005028100"
#>  [751] "36005028500" "36005028900" "36005029302" "36005029700" "36005030701"
#>  [756] "36005031900" "36005033500" "36005034300" "36005035100" "36005040304"
#>  [761] "36005040900" "36005035800" "36005036800" "36005037200" "36005037600"
#>  [766] "36005038000" "36005038600" "36005039000" "36005039400" "36005039800"
#>  [771] "36005040600" "36005041400" "36005042000" "36005042400" "36005042800"
#>  [776] "36005043400" "36005043600" "36005044400" "36005044901" "36005045101"
#>  [781] "36005045600" "36005046000" "36005011000" "36005013000" "36005013800"
#>  [786] "36005015200" "36005016000" "36005016400" "36005018400" "36005026400"
#>  [791] "36005026602" "36005027402" "36005030000" "36005046201" "36005050400"
#>  [796] "36005020000" "36005022403" "36005022800" "36005023200" "36005023800"
#>  [801] "36005024400" "36005024800" "36005025200" "36005025600" "36005028600"
#>  [806] "36005029600" "36005031200" "36005031600" "36005032400" "36005032800"
#>  [811] "36005033201" "36005033600" "36005034000" "36005034400" "36005035000"
#>  [816] "36005006000" "36005012300" "36005013300" "36005014100" "36005014701"
#>  [821] "36005014900" "36005015300" "36005015700" "36005016300" "36005016700"
#>  [826] "36005018500" "36005033400" "36005036100" "36005036501" "36005036700"
#>  [831] "36005036902" "36005037300" "36005038500" "36005038900" "36005039300"
#>  [836] "36005039700" "36005025300" "36005026100" "36005026500" "36005039901"
#>  [841] "36005040100" "36005040501" "36005040702" "36005041300" "36005041900"
#>  [846] "36005042300" "36005042901" "36005043100" "36005020501" "36005021301"
#>  [851] "36005021502" "36005022701" "36005022703" "36005022902" "36005023301"
#>  [856] "36005023501" "36005023703" "36005023900" "36005024300" "36005024502"
#>  [861] "36005024900" "36005025700" "36005038100" "36005038302" "36005006100"
#>  [866] "36005014300" "36005017300" "36005017701" "36005017901" "36005018101"
#>  [871] "36005018301" "36005018900" "36005019500" "36005019900" "36005020900"
#>  [876] "36005021302" "36005022101" "36005022300" "36005000200" "36005001600"
#>  [881] "36005002400" "36005003800" "36005004200" "36005004600" "36005005001"
#>  [886] "36005005200" "36005005600" "36005006400" "36005007000" "36005007400"
#>  [891] "36005007800" "36005008600" "36005009200" "36005009800" "36005020400"
#>  [896] "36005021001" "36005021200" "36005021602" "36005022200" "36005001900"
#>  [901] "36005002500" "36005002702" "36005003300" "36005003700" "36005004100"
#>  [906] "36005005100" "36005006700" "36005007100" "36005007500" "36005007900"
#>  [911] "36005008500" "36005008900" "36005011502" "36005011900" "36005012701"
#>  [916] "36005013100" "36007000100" "36007000300" "36007000500" "36007000700"
#>  [921] "36007001100" "36007001300" "36007001500" "36007001700" "36007012800"
#>  [926] "36007013100" "36007013202" "36007013304" "36007013500" "36007013700"
#>  [931] "36007013900" "36007014100" "36007014301" "36007014400" "36007010200"
#>  [936] "36007011902" "36007012000" "36007012102" "36007012201" "36007012600"
#>  [941] "36007012702" "36007013303" "36007012300" "36007012500" "36009940200"
#>  [946] "36009960100" "36009960300" "36009960500" "36009960702" "36009961000"
#>  [951] "36009961200" "36009961400" "36009961600" "36009961800" "36011040100"
#>  [956] "36011040300" "36011040500" "36011040700" "36011040900" "36011041100"
#>  [961] "36011041300" "36011041500" "36011041700" "36011042100" "36013030100"
#>  [966] "36013030300" "36013030500" "36013030700" "36013035100" "36013035400"
#>  [971] "36013035600" "36013035800" "36013035902" "36013036100" "36013036401"
#>  [976] "36013036500" "36013036700" "36013036901" "36013037000" "36013037200"
#>  [981] "36013037400" "36013037600" "36015000100" "36015000300" "36015000500"
#>  [986] "36015000700" "36015001000" "36015010300" "36015010500" "36015010700"
#>  [991] "36015010900" "36015011100" "36015010100" "36017970100" "36017970300"
#>  [996] "36017970500" "36017970602" "36017970801" "36017970900" "36019100100"
#> [1001] "36019100300" "36019100600" "36019100800" "36019101000" "36019101300"
#> [1006] "36019101600" "36019101800" "36019102000" "36019102200" "36021000200"
#> [1011] "36021000401" "36021000500" "36021000700" "36021000900" "36021001100"
#> [1016] "36021001300" "36021001500" "36021001700" "36021001900" "36023970100"
#> [1021] "36023970300" "36023970500" "36023970700" "36023970900" "36023971100"
#> [1026] "36025970100" "36025970300" "36025970500" "36025970700" "36025970900"
#> [1031] "36025971100" "36025971300" "36027010000" "36027020004" "36027030000"
#> [1036] "36027040003" "36027070102" "36027070301" "36027080103" "36027080201"
#> [1041] "36027090000" "36027110003" "36027110005" "36027130003" "36027130005"
#> [1046] "36027150004" "36027150006" "36027160004" "36027170000" "36027200001"
#> [1051] "36027610000" "36027050103" "36027050203" "36027050205" "36027060201"
#> [1056] "36027060301" "36027060400" "36027140200" "36027140400" "36027140602"
#> [1061] "36027140801" "36027190102" "36027190204" "36027190401" "36027210101"
#> [1066] "36027210301" "36027220201" "36027220700" "36027220901" "36027221100"
#> [1071] "36027410000" "36027640002" "36029007303" "36029007600" "36029007800"
#> [1076] "36029007902" "36029007904" "36029008001" "36029008003" "36029008102"
#> [1081] "36029008202" "36029008400" "36029008600" "36029008800" "36029008900"
#> [1086] "36029009006" "36029009008" "36029009010" "36029009106" "36029009109"
#> [1091] "36029009112" "36029009114" "36029009116" "36029009301" "36029009401"
#> [1096] "36029009501" "36029009600" "36029014102" "36029014206" "36029014208"
#> [1101] "36029014300" "36029014501" "36029014601" "36029014604" "36029014702"
#> [1106] "36029014803" "36029014903" "36029015001" "36029009701" "36029009800"
#> [1111] "36029010001" "36029010003" "36029010102" "36029010201" "36029010300"
#> [1116] "36029010500" "36029010700" "36029010804" "36029010807" "36029010809"
#> [1121] "36029010902" "36029011100" "36029011300" "36029011500" "36029011700"
#> [1126] "36029012001" "36029012003" "36029012400" "36029012502" "36029017300"
#> [1131] "36029000110" "36029000500" "36029000700" "36029000900" "36029001100"
#> [1136] "36029001500" "36029001700" "36029002300" "36029002502" "36029002800"
#> [1141] "36029003000" "36029003301" "36029003400" "36029003600" "36029003800"
#> [1146] "36029004001" "36029004200" "36029004401" "36029004602" "36029005202"
#> [1151] "36029016400" "36029016600" "36029016800" "36029004500" "36029004800"
#> [1156] "36029005000" "36029005201" "36029005400" "36029005600" "36029005801"
#> [1161] "36029005900" "36029006201" "36029006302" "36029006601" "36029006701"
#> [1166] "36029006800" "36029006902" "36029007101" "36029007202" "36029017100"
#> [1171] "36029012901" "36029013001" "36029013101" "36029013201" "36029013300"
#> [1176] "36029013501" "36029013600" "36029013702" "36029013900" "36029015002"
#> [1181] "36029015101" "36029015201" "36029015301" "36029015401" "36029015501"
#> [1186] "36029015504" "36029015700" "36029015900" "36029017501" "36029940000"
#> [1191] "36031960100" "36031960300" "36031960598" "36031960800" "36031961000"
#> [1196] "36031961200" "36031961400" "36033950100" "36033950300" "36033950501"
#> [1201] "36033950600" "36033950800" "36033951000" "36033951200" "36035970200"
#> [1206] "36035970400" "36035970600" "36035970800" "36035971000" "36035971200"
#> [1211] "36035971400" "36037940100" "36037950200" "36037950400" "36037950600"
#> [1216] "36037950800" "36037951000" "36037951200" "36037951400" "36039080201"
#> [1221] "36039080300" "36039080402" "36039080502" "36039080700" "36039080900"
#> [1226] "36039081101" "36041950100" "36041950400" "36043010201" "36043010300"
#> [1231] "36043010501" "36043010701" "36043010900" "36043011002" "36043011200"
#> [1236] "36043011302" "36043011501" "36043010100" "36045060200" "36045060400"
#> [1241] "36045060600" "36045060803" "36045060900" "36045061100" "36045061300"
#> [1246] "36045061500" "36045061700" "36045061900" "36045062200" "36045062500"
#> [1251] "36045990001" "36047047700" "36047049100" "36047049700" "36047050100"
#> [1256] "36047050500" "36047051100" "36047051500" "36047051900" "36047052500"
#> [1261] "36047052900" "36047053500" "36047053900" "36047054700" "36047055100"
#> [1266] "36047055500" "36047056100" "36047056500" "36047057100" "36047057500"
#> [1271] "36047058900" "36047059300" "36047038900" "36047039300" "36047039700"
#> [1276] "36047040100" "36047040500" "36047040900" "36047041300" "36047041700"
#> [1281] "36047042100" "36047042500" "36047042900" "36047043300" "36047043700"
#> [1286] "36047044100" "36047044500" "36047045300" "36047048900" "36047023300"
#> [1291] "36047024300" "36047024900" "36047025300" "36047025700" "36047025902"
#> [1296] "36047026300" "36047026700" "36047027300" "36047027700" "36047028100"
#> [1301] "36047028502" "36047028900" "36047029300" "36047037500" "36047037900"
#> [1306] "36047038500" "36047050700" "36047123700" "36047000301" "36047000502"
#> [1311] "36047000900" "36047001300" "36047002100" "36047002901" "36047003300"
#> [1316] "36047003700" "36047004100" "36047006900" "36047012700" "36047018100"
#> [1321] "36047018501" "36047019100" "36047019500" "36047019900" "36047021100"
#> [1326] "36047022900" "36047023500" "36047004500" "36047004900" "36047005300"
#> [1331] "36047006300" "36047006700" "36047007700" "36047011700" "36047012100"
#> [1336] "36047012902" "36047013300" "36047013700" "36047014100" "36047014900"
#> [1341] "36047015300" "36047015700" "36047016500" "36047017700" "36047016300"
#> [1346] "36047020500" "36047021500" "36047021900" "36047024700" "36047029700"
#> [1351] "36047030500" "36047030900" "36047031300" "36047031701" "36047033700"
#> [1356] "36047034100" "36047034500" "36047034900" "36047035300" "36047035900"
#> [1361] "36047030100" "36047036100" "36047036501" "36047036700" "36047037100"
#> [1366] "36047089200" "36047089600" "36047090000" "36047090600" "36047091000"
#> [1371] "36047091600" "36047092000" "36047092400" "36047112600" "36047113000"
#> [1376] "36047113400" "36047115600" "36047105801" "36047107000" "36047109800"
#> [1381] "36047110600" "36047111600" "36047112000" "36047114201" "36047114600"
#> [1386] "36047115200" "36047116200" "36047116600" "36047117000" "36047117202"
#> [1391] "36047117601" "36047117800" "36047118201" "36047118400" "36047118800"
#> [1396] "36047119200" "36047119600" "36047120000" "36047120800" "36047121400"
#> [1401] "36047063600" "36047064400" "36047064800" "36047065200" "36047065600"
#> [1406] "36047066000" "36047066600" "36047067200" "36047067600" "36047068000"
#> [1411] "36047068600" "36047069000" "36047069601" "36047069800" "36047070201"
#> [1416] "36047070203" "36047072000" "36047072400" "36047072800" "36047073200"
#> [1421] "36047073600" "36047074000" "36047074400" "36047077600" "36047094402"
#> [1426] "36047095400" "36047095800" "36047096200" "36047096600" "36047097000"
#> [1431] "36047098200" "36047098600" "36047099000" "36047099400" "36047099800"
#> [1436] "36047100600" "36047101000" "36047101400" "36047101800" "36047102200"
#> [1441] "36047102600" "36047103400" "36047078200" "36047079000" "36047079400"
#> [1446] "36047081600" "36047082400" "36047082800" "36047083200" "36047083600"
#> [1451] "36047084000" "36047084800" "36047085200" "36047085600" "36047086000"
#> [1456] "36047086400" "36047086800" "36047087200" "36047088400" "36047088800"
#> [1461] "36047092800" "36047093200" "36047093600" "36047094600" "36047031900"
#> [1466] "36047032300" "36047032700" "36047033100" "36047033500" "36047079601"
#> [1471] "36047079801" "36047080000" "36047080400" "36047080800" "36047082000"
#> [1476] "36047087401" "36047087800" "36047000200" "36047002000" "36047007200"
#> [1481] "36047007600" "36047008000" "36047008400" "36047008800" "36047009200"
#> [1486] "36047009600" "36047010000" "36047010200" "36047010600" "36047011000"
#> [1491] "36047011800" "36047012200" "36047014700" "36047017100" "36047050000"
#> [1496] "36047050400" "36047003000" "36047003600" "36047004400" "36047005000"
#> [1501] "36047005202" "36047005601" "36047005800" "36047006200" "36047006600"
#> [1506] "36047007000" "36047012801" "36047013200" "36047013600" "36047014000"
#> [1511] "36047014800" "36047015200" "36047016000" "36047016400" "36047019400"
#> [1516] "36047019800" "36047020200" "36047020600" "36047021000" "36047011400"
#> [1521] "36047019200" "36047021600" "36047022000" "36047022400" "36047022800"
#> [1526] "36047023200" "36047023600" "36047024000" "36047024400" "36047043800"
#> [1531] "36047044200" "36047044600" "36047045000" "36047045400" "36047046400"
#> [1536] "36047047000" "36047047400" "36047047800" "36047048600" "36047049000"
#> [1541] "36047049400" "36047049800" "36047045800" "36047046201" "36047048200"
#> [1546] "36047050801" "36047050804" "36047051002" "36047051400" "36047051602"
#> [1551] "36047052000" "36047052800" "36047053200" "36047053800" "36047054400"
#> [1556] "36047074800" "36047075200" "36047075600" "36047076000" "36047076400"
#> [1561] "36047076800" "36047077200" "36047078600" "36047152200" "36047039000"
#> [1566] "36047039400" "36047041401" "36047041600" "36047042000" "36047054800"
#> [1571] "36047055200" "36047055600" "36047056000" "36047056400" "36047056800"
#> [1576] "36047057200" "36047057600" "36047058000" "36047058400" "36047058800"
#> [1581] "36047059200" "36047059402" "36047059800" "36047060600" "36047061200"
#> [1586] "36047062000" "36047062600" "36047063200" "36047064200" "36047017000"
#> [1591] "36047017400" "36047017800" "36047018200" "36047018600" "36047019000"
#> [1596] "36047025000" "36047025400" "36047025800" "36047026200" "36047026600"
#> [1601] "36047027000" "36047027400" "36047027800" "36047028200" "36047028600"
#> [1606] "36047029000" "36047029400" "36047029800" "36047030200" "36047040000"
#> [1611] "36047040400" "36047040800" "36047041200" "36047042600" "36047043000"
#> [1616] "36047043400" "36047030600" "36047031400" "36047032800" "36047033600"
#> [1621] "36047034200" "36047035000" "36047035400" "36047035602" "36047036002"
#> [1626] "36047036400" "36047037000" "36047037402" "36047038600" "36047061002"
#> [1631] "36047061004" "36049950100" "36049950300" "36049950500" "36049950700"
#> [1636] "36051030201" "36051030300" "36051030500" "36051030700" "36051030900"
#> [1641] "36051031100" "36051031300" "36053030501" "36053030700" "36053030900"
#> [1646] "36053031100" "36053940102" "36053940200" "36053940401" "36053940600"
#> [1651] "36055010100" "36055010300" "36055010500" "36055010602" "36055010800"
#> [1656] "36055010902" "36055011100" "36055011203" "36055011207" "36055011301"
#> [1661] "36055011400" "36055011503" "36055011505" "36055011603" "36055011605"
#> [1666] "36055000700" "36055001300" "36055002900" "36055003100" "36055003300"
#> [1671] "36055003500" "36055003700" "36055004602" "36055004702" "36055004900"
#> [1676] "36055005100" "36055005300" "36055005500" "36055005700" "36055005900"
#> [1681] "36055006100" "36055007700" "36055007802" "36055008000" "36055008200"
#> [1686] "36055008400" "36055009301" "36055000200" "36055001900" "36055002100"
#> [1691] "36055002300" "36055002700" "36055003805" "36055004100" "36055006300"
#> [1696] "36055006500" "36055006700" "36055006900" "36055007100" "36055008500"
#> [1701] "36055008701" "36055008800" "36055009500" "36055009602" "36055009604"
#> [1706] "36055013401" "36055013503" "36055013506" "36055013508" "36055013603"
#> [1711] "36055013701" "36055013800" "36055013902" "36055014003" "36055014102"
#> [1716] "36055014104" "36055014203" "36055014301" "36055014400" "36055014503"
#> [1721] "36055014505" "36055014602" "36055014802" "36055014804" "36055014903"
#> [1726] "36055014906" "36055015101" "36055015200" "36055015303" "36055015400"
#> [1731] "36055011703" "36055011706" "36055011708" "36055011901" "36055012000"
#> [1736] "36055012201" "36055012301" "36055012305" "36055012401" "36055012500"
#> [1741] "36055012700" "36055012900" "36055013002" "36055013103" "36055013203"
#> [1746] "36055013205" "36055013300" "36057070300" "36057070500" "36057070700"
#> [1751] "36057070900" "36057072200" "36057072400" "36057072600" "36057072800"
#> [1756] "36059300300" "36059300500" "36059300700" "36059300900" "36059301101"
#> [1761] "36059301200" "36059301400" "36059301600" "36059301800" "36059302000"
#> [1766] "36059302102" "36059302502" "36059517000" "36059517102" "36059517301"
#> [1771] "36059517400" "36059517600" "36059517705" "36059517802" "36059517902"
#> [1776] "36059518100" "36059518203" "36059518300" "36059518501" "36059981100"
#> [1781] "36059990200" "36059518502" "36059518800" "36059519000" "36059519200"
#> [1786] "36059519400" "36059519601" "36059519702" "36059519704" "36059519802"
#> [1791] "36059520001" "36059520100" "36059302400" "36059302600" "36059302800"
#> [1796] "36059303000" "36059303102" "36059303202" "36059303302" "36059303500"
#> [1801] "36059303700" "36059303900" "36059304002" "36059304202" "36059304204"
#> [1806] "36059404400" "36059404600" "36059404800" "36059404902" "36059405100"
#> [1811] "36059405301" "36059405400" "36059405600" "36059405800" "36059406400"
#> [1816] "36059406600" "36059406701" "36059406801" "36059406900" "36059407101"
#> [1821] "36059407201" "36059407204" "36059407302" "36059407402" "36059407502"
#> [1826] "36059414001" "36059414100" "36059414202" "36059414303" "36059414400"
#> [1831] "36059414502" "36059407700" "36059407802" "36059408000" "36059408200"
#> [1836] "36059408600" "36059408800" "36059409000" "36059409200" "36059409400"
#> [1841] "36059409600" "36059415500" "36059520300" "36059520402" "36059520502"
#> [1846] "36059520700" "36059520900" "36059521100" "36059521301" "36059521400"
#> [1851] "36059521601" "36059521700" "36059521802" "36059522000" "36059990400"
#> [1856] "36059408500" "36059414700" "36059414900" "36059415101" "36059415201"
#> [1861] "36059415300" "36059415402" "36059415700" "36059416000" "36059522700"
#> [1866] "36059412302" "36059412500" "36059412700" "36059412900" "36059413002"
#> [1871] "36059413200" "36059413400" "36059413600" "36059413803" "36059416201"
#> [1876] "36059416900" "36059406001" "36059406100" "36059406202" "36059409900"
#> [1881] "36059410100" "36059410300" "36059410500" "36059410700" "36059411800"
#> [1886] "36059411902" "36059410900" "36059411100" "36059411301" "36059411400"
#> [1891] "36059411600" "36059412100" "36059416300" "36059416402" "36059416600"
#> [1896] "36059416702" "36059416802" "36061024100" "36061024500" "36061024900"
#> [1901] "36061025300" "36061026100" "36061026500" "36061026900" "36061027300"
#> [1906] "36061027700" "36061028100" "36061028500" "36061029100" "36061029500"
#> [1911] "36061029900" "36061030700" "36061031100" "36061019500" "36061019900"
#> [1916] "36061020300" "36061020701" "36061021100" "36061021703" "36061022301"
#> [1921] "36061022500" "36061022900" "36061023300" "36061023700" "36061019000"
#> [1926] "36061020000" "36061020600" "36061021200" "36061021500" "36061021800"
#> [1931] "36061022102" "36061022400" "36061022800" "36061023200" "36061023502"
#> [1936] "36061024302" "36061025900" "36061015802" "36061016200" "36061016600"
#> [1941] "36061017000" "36061017401" "36061017800" "36061018200" "36061018800"
#> [1946] "36061019400" "36061019800" "36061024000" "36061010602" "36061011401"
#> [1951] "36061011600" "36061012000" "36061012400" "36061012800" "36061013200"
#> [1956] "36061013600" "36061014000" "36061014401" "36061014601" "36061014801"
#> [1961] "36061015001" "36061015200" "36061015601" "36061016001" "36061023802"
#> [1966] "36061014500" "36061014900" "36061015300" "36061015700" "36061016100"
#> [1971] "36061016500" "36061016900" "36061017300" "36061017700" "36061018100"
#> [1976] "36061018500" "36061018900" "36061005200" "36061005600" "36061007400"
#> [1981] "36061008100" "36061008400" "36061008900" "36061009300" "36061009500"
#> [1986] "36061009700" "36061010100" "36061010300" "36061010900" "36061011201"
#> [1991] "36061011300" "36061011700" "36061012100" "36061012700" "36061013100"
#> [1996] "36061013500" "36061013900" "36061004800" "36061006000" "36061006400"
#> [2001] "36061006800" "36061007200" "36061008000" "36061008601" "36061008603"
#> [2006] "36061009000" "36061009800" "36061010601" "36061011203" "36061000202"
#> [2011] "36061000800" "36061001002" "36061001401" "36061001600" "36061002000"
#> [2016] "36061002202" "36061002500" "36061002602" "36061002800" "36061003001"
#> [2021] "36061003200" "36061003601" "36061003800" "36061004200" "36061000500"
#> [2026] "36061000900" "36061001501" "36061002100" "36061003300" "36061003900"
#> [2031] "36061004300" "36061004700" "36061005501" "36061005700" "36061006100"
#> [2036] "36061006500" "36061006900" "36061007300" "36061007700" "36061031703"
#> [2041] "36061031900" "36063020200" "36063020400" "36063020600" "36063020900"
#> [2046] "36063021100" "36063021300" "36063021700" "36063022100" "36063022300"
#> [2051] "36063022500" "36063022602" "36063022712" "36063022804" "36063022902"
#> [2056] "36063023100" "36063023300" "36063022702" "36063023402" "36063023405"
#> [2061] "36063023600" "36063023800" "36063023902" "36063024002" "36063024102"
#> [2066] "36063024202" "36063024302" "36063024401" "36063024405" "36063024501"
#> [2071] "36063940001" "36063990000" "36065023902" "36065024101" "36065024200"
#> [2076] "36065024302" "36065024400" "36065026700" "36065940100" "36065020102"
#> [2081] "36065020705" "36065020803" "36065021000" "36065021102" "36065021201"
#> [2086] "36065021301" "36065021303" "36065021402" "36065021601" "36065021701"
#> [2091] "36065021900" "36065022100" "36065022400" "36065022701" "36065022800"
#> [2096] "36065023200" "36065023400" "36065023502" "36065025001" "36065025003"
#> [2101] "36065025200" "36065025400" "36065026100" "36065026300" "36065026500"
#> [2106] "36065980002" "36065980100" "36065025600" "36065025800" "36067000200"
#> [2111] "36067000400" "36067000600" "36067000800" "36067001000" "36067001500"
#> [2116] "36067001701" "36067001800" "36067002000" "36067002300" "36067002700"
#> [2121] "36067003000" "36067003400" "36067003601" "36067003800" "36067004000"
#> [2126] "36067004301" "36067004400" "36067004600" "36067004900" "36067005100"
#> [2131] "36067005300" "36067005500" "36067005602" "36067005800" "36067006000"
#> [2136] "36067006102" "36067010100" "36067010301" "36067010322" "36067010500"
#> [2141] "36067011012" "36067011022" "36067011102" "36067011202" "36067011232"
#> [2146] "36067011242" "36067011401" "36067011500" "36067011700" "36067011900"
#> [2151] "36067013500" "36067010700" "36067010900" "36067012200" "36067012400"
#> [2156] "36067012600" "36067012800" "36067013000" "36067013200" "36067013600"
#> [2161] "36067013800" "36067014000" "36067014300" "36067014500" "36067014700"
#> [2166] "36067014900" "36067015100" "36067015202" "36067015400" "36067015601"
#> [2171] "36067016100" "36067016300" "36067016502" "36067012000" "36067015800"
#> [2176] "36067016501" "36067016801" "36067016901" "36067940000" "36069050102"
#> [2181] "36069050202" "36069050302" "36069050500" "36069050602" "36069050900"
#> [2186] "36069051100" "36069051300" "36069051500" "36069051700" "36069051900"
#> [2191] "36069052100" "36071000100" "36071000300" "36071000501" "36071000600"
#> [2196] "36071010102" "36071010300" "36071010500" "36071010700" "36071010802"
#> [2201] "36071012602" "36071012800" "36071013000" "36071015200" "36071001200"
#> [2206] "36071001500" "36071002100" "36071002300" "36071010902" "36071011101"
#> [2211] "36071011200" "36071011400" "36071011601" "36071011701" "36071011801"
#> [2216] "36071011900" "36071012200" "36071015100" "36071013202" "36071013400"
#> [2221] "36071013600" "36071013800" "36071014101" "36071014201" "36071014301"
#> [2226] "36071014400" "36071014502" "36071014700" "36071014900" "36071015004"
#> [2231] "36071015006" "36073040300" "36073040500" "36073040700" "36073040802"
#> [2236] "36073401300" "36075020100" "36075020301" "36075020400" "36075020600"
#> [2241] "36075020702" "36075020800" "36075020902" "36075021101" "36075021103"
#> [2246] "36075021200" "36075021401" "36075021501" "36075021601" "36075021603"
#> [2251] "36075021605" "36077590100" "36077590202" "36077590400" "36077590600"
#> [2256] "36077590800" "36077591000" "36077591200" "36077591400" "36077591600"
#> [2261] "36079010200" "36079010400" "36079010600" "36079010800" "36079011000"
#> [2266] "36079011200" "36079011400" "36079011600" "36079011800" "36081002500"
#> [2271] "36081003300" "36081003900" "36081004500" "36081005100" "36081005500"
#> [2276] "36081005900" "36081006300" "36081006502" "36081007100" "36081007500"
#> [2281] "36081007900" "36081008300" "36081008700" "36081009500" "36081009900"
#> [2286] "36081010300" "36081010701" "36081011300" "36081011700" "36081012100"
#> [2291] "36081012500" "36081013700" "36081014300" "36081014700" "36081015100"
#> [2296] "36081015500" "36081015900" "36081016300" "36081031700" "36081027500"
#> [2301] "36081027900" "36081028300" "36081028700" "36081029100" "36081030903"
#> [2306] "36081032700" "36081033100" "36081033900" "36081035100" "36081035700"
#> [2311] "36081036300" "36081036700" "36081037300" "36081037700" "36081038100"
#> [2316] "36081040300" "36081040700" "36081038301" "36081079702" "36081080301"
#> [2321] "36081083700" "36081084900" "36081085500" "36081085900" "36081086300"
#> [2326] "36081086900" "36081088901" "36081091900" "36081092900" "36081094500"
#> [2331] "36081097300" "36081098700" "36081099701" "36081099704" "36081099900"
#> [2336] "36081102900" "36081103900" "36081105900" "36081114700" "36081115500"
#> [2341] "36081115900" "36081116300" "36081117100" "36081118500" "36081118900"
#> [2346] "36081119300" "36081119900" "36081120300" "36081120700" "36081121500"
#> [2351] "36081109300" "36081109900" "36081112300" "36081113300" "36081118100"
#> [2356] "36081129103" "36081136700" "36081138501" "36081139900" "36081140901"
#> [2361] "36081141700" "36081143500" "36081144700" "36081145102" "36081146300"
#> [2366] "36081147100" "36081148300" "36081150702" "36081152902" "36081032000"
#> [2371] "36081035800" "36081051200" "36081053200" "36081053601" "36081054000"
#> [2376] "36081054800" "36081055400" "36081055800" "36081056200" "36081056600"
#> [2381] "36081058000" "36081059000" "36081059400" "36081059800" "36081060600"
#> [2386] "36081061000" "36081061400" "36081061602" "36081062000" "36081062400"
#> [2391] "36081063000" "36081063800" "36081065000" "36081065600" "36081066400"
#> [2396] "36081068200" "36081069400" "36081130100" "36081155102" "36081157101"
#> [2401] "36081157901" "36081157903" "36081162100" "36081022001" "36081023000"
#> [2406] "36081023600" "36081045000" "36081045400" "36081045800" "36081046600"
#> [2411] "36081047600" "36081049200" "36081077903" "36081077905" "36081077907"
#> [2416] "36081079300" "36081122300" "36081122702" "36081124700" "36081126500"
#> [2421] "36081127700" "36081133300" "36081134100" "36081026700" "36081026902"
#> [2426] "36081038302" "36081041100" "36081041500" "36081043701" "36081043900"
#> [2431] "36081044302" "36081045700" "36081046100" "36081046500" "36081046900"
#> [2436] "36081047300" "36081048100" "36081068300" "36081068700" "36081069500"
#> [2441] "36081069702" "36081070700" "36081071100" "36081071304" "36081071306"
#> [2446] "36081071702" "36081072100" "36081072900" "36081073700" "36081074100"
#> [2451] "36081074500" "36081074900" "36081075702" "36081076902" "36081000700"
#> [2456] "36081016900" "36081017900" "36081018102" "36081018501" "36081018700"
#> [2461] "36081019900" "36081021900" "36081023500" "36081024500" "36081024900"
#> [2466] "36081025301" "36081025500" "36081025900" "36081026300" "36081029300"
#> [2471] "36081029700" "36081048300" "36081048900" "36081049302" "36081049700"
#> [2476] "36081050700" "36081051300" "36081051700" "36081052500" "36081053500"
#> [2481] "36081054500" "36081054900" "36081055300" "36081055700" "36081056100"
#> [2486] "36081056700" "36081057900" "36081058300" "36081058700" "36081059100"
#> [2491] "36081059500" "36081060100" "36081060701" "36081061302" "36081062100"
#> [2496] "36081062500" "36081062900" "36081063302" "36081063700" "36081065501"
#> [2501] "36081065703" "36081066100" "36081066501" "36081066900" "36081067700"
#> [2506] "36081000200" "36081000600" "36081001000" "36081001400" "36081001800"
#> [2511] "36081002200" "36081002600" "36081003000" "36081003400" "36081003800"
#> [2516] "36081004200" "36081010800" "36081011200" "36081011600" "36081012000"
#> [2521] "36081012400" "36081012602" "36081013000" "36081013400" "36081013800"
#> [2526] "36081014201" "36081014400" "36081015000" "36081015400" "36081021600"
#> [2531] "36081064102" "36081077500" "36081018401" "36081018600" "36081019000"
#> [2536] "36081019400" "36081019800" "36081020400" "36081020800" "36081023800"
#> [2541] "36081024600" "36081025800" "36081026200" "36081026600" "36081027200"
#> [2546] "36081027600" "36081028000" "36081028400" "36081029400" "36081033401"
#> [2551] "36081035200" "36081036800" "36081038400" "36081039800" "36081040200"
#> [2556] "36081041400" "36081042600" "36081043400" "36081044400" "36081044602"
#> [2561] "36081046200" "36081047000" "36081048200" "36081050000" "36081050202"
#> [2566] "36081050600" "36081051000" "36081052000" "36081052400" "36081052800"
#> [2571] "36081078800" "36081079200" "36081004401" "36081005400" "36081006201"
#> [2576] "36081008600" "36081009400" "36081009800" "36081010200" "36081010600"
#> [2581] "36081015802" "36081016600" "36081017000" "36081017400" "36081017800"
#> [2586] "36081081400" "36081083800" "36081084601" "36081086400" "36081089200"
#> [2591] "36081091602" "36081092200" "36081093401" "36081093800" "36081094202"
#> [2596] "36081095400" "36081097202" "36081097204" "36081099801" "36081100801"
#> [2601] "36081101001" "36081103201" "36081107201" "36081990100" "36083040200"
#> [2606] "36083040400" "36083040600" "36083040800" "36083041000" "36083041200"
#> [2611] "36083041400" "36083051600" "36083051702" "36083051901" "36083052002"
#> [2616] "36083052004" "36083052102" "36083052201" "36083052204" "36083052303"
#> [2621] "36083052402" "36083052404" "36083052502" "36083052601" "36083052603"
#> [2626] "36085012806" "36085013203" "36085013800" "36085014605" "36085014607"
#> [2631] "36085015400" "36085015602" "36085017005" "36085017008" "36085017010"
#> [2636] "36085017012" "36085019800" "36085020803" "36085022600" "36085024401"
#> [2641] "36085024800" "36085001800" "36085005000" "36085007000" "36085009601"
#> [2646] "36085011201" "36085011401" "36085012200" "36085013400" "36085017700"
#> [2651] "36085018702" "36085027301" "36085027702" "36085027705" "36085027900"
#> [2656] "36085029103" "36085000300" "36085000700" "36085000900" "36085001700"
#> [2661] "36085002100" "36085002900" "36085003600" "36085004000" "36085005900"
#> [2666] "36085007500" "36085008100" "36085009700" "36085012100" "36085013301"
#> [2671] "36085014100" "36085015100" "36085018701" "36085019700" "36085020700"
#> [2676] "36085022300" "36085023900" "36085025100" "36085030302" "36085031902"
#> [2681] "36087010101" "36087010200" "36087010502" "36087010601" "36087010701"
#> [2686] "36087010703" "36087010802" "36087010804" "36087010902" "36087011401"
#> [2691] "36087011404" "36087011000" "36087011200" "36087011302" "36087011602"
#> [2696] "36087012501" "36087012600" "36087012800" "36087013002" "36087013100"
#> [2701] "36087013300" "36087013402" "36087011502" "36087011505" "36087011601"
#> [2706] "36087011800" "36087011902" "36087012101" "36087012103" "36087012106"
#> [2711] "36087012203" "36087012300" "36087012402" "36089490200" "36089490400"
#> [2716] "36089490600" "36089490800" "36089491000" "36089491200" "36089491400"
#> [2721] "36089491600" "36089491800" "36089492000" "36089492300" "36089492500"
#> [2726] "36089492700" "36089492900" "36091061100" "36091061301" "36091061303"
#> [2731] "36091061800" "36091061903" "36091062404" "36091062406" "36091062503"
#> [2736] "36091062507" "36091062509" "36091062800" "36091060102" "36091060300"
#> [2741] "36091060501" "36091060503" "36091060602" "36091060702" "36091060901"
#> [2746] "36091061401" "36091061404" "36091061600" "36091062000" "36091062200"
#> [2751] "36091062505" "36091062602" "36093020102" "36093020300" "36093020600"
#> [2756] "36093020800" "36093021001" "36093021200" "36093021500" "36093021700"
#> [2761] "36093031900" "36093032101" "36093032200" "36093032402" "36093032404"
#> [2766] "36093032503" "36093032601" "36093032700" "36093032902" "36093033003"
#> [2771] "36093033101" "36093033200" "36093033400" "36095740100" "36095740300"
#> [2776] "36095740500" "36095740700" "36097950100" "36097950300" "36097950500"
#> [2781] "36099950200" "36099950400" "36099950600" "36099950800" "36099951000"
#> [2786] "36101962400" "36101962600" "36101962800" "36101960100" "36101960300"
#> [2791] "36101960500" "36101960700" "36101960900" "36101961100" "36101961300"
#> [2796] "36101961500" "36101961700" "36101961900" "36101962100" "36101963000"
#> [2801] "36103110102" "36103110300" "36103110402" "36103110502" "36103110801"
#> [2806] "36103110901" "36103111001" "36103111100" "36103111202" "36103111401"
#> [2811] "36103111505" "36103111602" "36103111504" "36103111701" "36103111704"
#> [2816] "36103111802" "36103111804" "36103112001" "36103112102" "36103112104"
#> [2821] "36103112206" "36103112211" "36103112213" "36103134702" "36103134704"
#> [2826] "36103134903" "36103134906" "36103135002" "36103135004" "36103135101"
#> [2831] "36103135103" "36103135201" "36103135205" "36103135209" "36103135303"
#> [2836] "36103135401" "36103135403" "36103158002" "36103158007" "36103158010"
#> [2841] "36103158202" "36103158205" "36103158207" "36103158306" "36103158318"
#> [2846] "36103158323" "36103158402" "36103158407" "36103158409" "36103169701"
#> [2851] "36103169704" "36103169901" "36103170001" "36103170101" "36103170202"
#> [2856] "36103190401" "36103190403" "36103190503" "36103190601" "36103190604"
#> [2861] "36103190705" "36103190707" "36103190800" "36103200902" "36103201003"
#> [2866] "36103990100" "36103158310" "36103158707" "36103158709" "36103159103"
#> [2871] "36103159204" "36103159404" "36103159407" "36103159410" "36103159412"
#> [2876] "36103159506" "36103159509" "36103159511" "36103159601" "36103158116"
#> [2881] "36103158319" "36103158321" "36103158607" "36103158705" "36103158712"
#> [2886] "36103158803" "36103158900" "36103159102" "36103159107" "36103159201"
#> [2891] "36103158102" "36103158104" "36103158108" "36103158111" "36103158114"
#> [2896] "36103158308" "36103158505" "36103158507" "36103158509" "36103158512"
#> [2901] "36103158605" "36103158608" "36103145701" "36103145805" "36103145808"
#> [2906] "36103146605" "36103146607" "36103146611" "36103146613" "36103146615"
#> [2911] "36103147701" "36103147802" "36103147804" "36103147902" "36103145603"
#> [2916] "36103145605" "36103145703" "36103145804" "36103145902" "36103146002"
#> [2921] "36103146105" "36103146201" "36103146203" "36103146300" "36103146403"
#> [2926] "36103146703" "36103145903" "36103146103" "36103146206" "36103146704"
#> [2931] "36103146706" "36103146901" "36103147001" "36103147004" "36103147200"
#> [2936] "36103147401" "36103147501" "36103147503" "36103122404" "36103122603"
#> [2941] "36103122901" "36103123001" "36103123101" "36103123900" "36103124002"
#> [2946] "36103124102" "36103124300" "36103124402" "36103124601" "36103122300"
#> [2951] "36103122405" "36103122501" "36103122602" "36103122705" "36103122707"
#> [2956] "36103123201" "36103123301" "36103123401" "36103123500" "36103123701"
#> [2961] "36103123801" "36105950100" "36105950300" "36105950500" "36105950700"
#> [2966] "36105950900" "36105951100" "36105951300" "36105951600" "36105951800"
#> [2971] "36105952000" "36105952200" "36105952400" "36107020100" "36107020300"
#> [2976] "36107020402" "36107020600" "36107020702" "36109000100" "36109000300"
#> [2981] "36109000500" "36109000700" "36109000900" "36109001100" "36109001300"
#> [2986] "36109001500" "36109001700" "36109001900" "36109002100" "36109002300"
#> [2991] "36111950600" "36111951000" "36111954500" "36111954700" "36111955000"
#> [2996] "36111950100" "36111950300" "36111951100" "36111951300" "36111951500"
#> [3001] "36111951700" "36111951900" "36111952100" "36111952300" "36111952500"
#> [3006] "36111952700" "36111952900" "36111953400" "36111953600" "36111953800"
#> [3011] "36111954000" "36111954200" "36111954900" "36113070100" "36113070300"
#> [3016] "36113070500" "36113070602" "36113070702" "36113070900" "36113072000"
#> [3021] "36113073500" "36113075000" "36113078000" "36115080200" "36115081000"
#> [3026] "36115082002" "36115085000" "36115087000" "36115089000" "36115091000"
#> [3031] "36115093000" "36117020101" "36117020201" "36117020301" "36117020401"
#> [3036] "36117020500" "36117020700" "36117020900" "36117021100" "36117021400"
#> [3041] "36117021502" "36117021700" "36117990100" "36119011600" "36119011800"
#> [3046] "36119012000" "36119012102" "36119013201" "36119013301" "36119013400"
#> [3051] "36119013600" "36119013800" "36119014000" "36119014200" "36119014400"
#> [3056] "36119014604" "36119014606" "36119014701" "36119014704" "36119982000"
#> [3061] "36119012301" "36119012304" "36119012501" "36119012503" "36119012700"
#> [3066] "36119012900" "36119013102" "36119013104" "36119014805" "36119014808"
#> [3071] "36119014810" "36119014901" "36119014907" "36119015000" "36119006900"
#> [3076] "36119007100" "36119007300" "36119007402" "36119007600" "36119007800"
#> [3081] "36119008000" "36119008200" "36119008302" "36119008403" "36119008500"
#> [3086] "36119008700" "36119008800" "36119008902" "36119009100" "36119009300"
#> [3091] "36119009500" "36119009701" "36119009703" "36119009900" "36119010100"
#> [3096] "36119010300" "36119010500" "36119010701" "36119010801" "36119010804"
#> [3101] "36119010902" "36119011000" "36119011102" "36119011300" "36119011500"
#> [3106] "36119000103" "36119000201" "36119000203" "36119000401" "36119000500"
#> [3111] "36119000701" "36119000801" "36119000803" "36119001000" "36119001102"
#> [3116] "36119001301" "36119001303" "36119001402" "36119001502" "36119001504"
#> [3121] "36119001600" "36119001800" "36119002000" "36119002103" "36119002105"
#> [3126] "36119002107" "36119002202" "36119002204" "36119002401" "36119002403"
#> [3131] "36119002405" "36119002700" "36119002900" "36119003100" "36119003300"
#> [3136] "36119003500" "36119003700" "36119003900" "36119004100" "36119004300"
#> [3141] "36119004500" "36119004700" "36119004900" "36119005002" "36119005200"
#> [3146] "36119005400" "36119005600" "36119005702" "36119005901" "36119006000"
#> [3151] "36119006200" "36119006400" "36119006600" "36119006801" "36121970100"
#> [3156] "36121970300" "36121970500" "36121970700" "36121970900" "36121971100"
#> [3161] "36123150200" "36123150400"