This function filters the cbsa dataframe.

This function takes in a cbsa dataset and a dynamic list of user selected columns. It returns a filtered dataframes based on user conditions

Usage

preview_meta(dataset, visual, within)

Arguments

dataset

string. Name of dataset to load, "cbsa", "block", "tract".

visual

boolean. Option to return reactable visualization or filtered dataframe

within

character vector. Vector of column variables to filter by without explicit argument definition, filters all rows with columns containing any of the within arguments. For example within = c("NY", "Alabama") will return all rows containing either NY or Alabama

...

expression. User inputs of selected columns and values to filter by. E.g. (state.census.abbr = c("NY", "AL")). Leaving blank returns all columns

Value

filtered dataframe or list with filtered dataframe and first 20 rows of table visualized with reactable

Examples

preview_meta("cbsa", TRUE, state.census.name = c("Wyoming", "Montana"))
#> [[1]]
#> 
#> [[2]]
#>    metro.census.cbsa.geoid metro.div.geoid metro.census.csa.geoid
#> 1                    13740              NA                     NA
#> 2                    28060              NA                     NA
#> 3                    25740              NA                     NA
#> 4                    33540              NA                     NA
#> 5                    15580              NA                     NA
#> 6                    13740              NA                     NA
#> 7                    29660              NA                     NA
#> 8                    40180              NA                     NA
#> 9                    16220              NA                     NA
#> 10                   43260              NA                     NA
#> 11                   27220              NA                     NA
#>    metro.census.cbsa.name              metro.micro.name metro.div.name
#> 1            Billings, MT Metropolitan Statistical Area               
#> 2           Kalispell, MT Micropolitan Statistical Area               
#> 3              Helena, MT Metropolitan Statistical Area               
#> 4            Missoula, MT Metropolitan Statistical Area               
#> 5    Butte-Silver Bow, MT Micropolitan Statistical Area               
#> 6            Billings, MT Metropolitan Statistical Area               
#> 7             Laramie, WY Micropolitan Statistical Area               
#> 8            Riverton, WY Micropolitan Statistical Area               
#> 9              Casper, WY Metropolitan Statistical Area               
#> 10           Sheridan, WY Micropolitan Statistical Area               
#> 11         Jackson, WY-ID Micropolitan Statistical Area               
#>    metro.census.csa.name census.county.name state.census.name state.censu.geoid
#> 1                             Carbon County           Montana                30
#> 2                           Flathead County           Montana                30
#> 3                          Jefferson County           Montana                30
#> 4                            Mineral County           Montana                30
#> 5                         Silver Bow County           Montana                30
#> 6                        Yellowstone County           Montana                30
#> 7                             Albany County           Wyoming                56
#> 8                            Fremont County           Wyoming                56
#> 9                            Natrona County           Wyoming                56
#> 10                          Sheridan County           Wyoming                56
#> 11                             Teton County           Wyoming                56
#>    census.county.geoid census.centrout.name state.census.abbr
#> 1                    9             Outlying                MT
#> 2                   29              Central                MT
#> 3                   43             Outlying                MT
#> 4                   61             Outlying                MT
#> 5                   93              Central                MT
#> 6                  111              Central                MT
#> 7                    1              Central                WY
#> 8                   13              Central                WY
#> 9                   25              Central                WY
#> 10                  33              Central                WY
#> 11                  39              Central                WY
#> 
preview_meta("tract", TRUE,
                metro.census.cbsa.geoid = c("10100", "10200"),
                state.census.abbr = c("NY", "CA"))
#> [[1]]
#> 
#> [[2]]
#> Empty data.table (0 rows and 14 cols): tract.census.geoid,county.census.geoid,puma.census.geoid,state.census.geoid,state.census.name,metro.census.cbsa.geoid...
#> 
preview_meta("tract", TRUE, within = c("NY", "Alabama"))
#> [[1]]
#> 
#> [[2]]
#>    metro.census.cbsa.geoid metro.div.geoid metro.census.csa.geoid
#> 1                    33860              NA                    388
#> 2                    19300              NA                    380
#> 3                    21640              NA                     NA
#> 4                    13820              NA                    142
#> 5                    13820              NA                    142
#> 6                    11500              NA                     NA
#> 7                    29300              NA                    122
#> 8                    13820              NA                    142
#> 9                    21460              NA                    222
#> 10                   22520              NA                    250
#> 11                   45180              NA                    142
#> 12                   18980              NA                    142
#> 13                   37120              NA                    222
#> 14                   42820              NA                    388
#> 15                   22840              NA                    290
#> 16                   33860              NA                    388
#> 17                   23460              NA                     NA
#> 18                   40770              NA                    250
#> 19                   20020              NA                    222
#> 20                   46220              NA                     NA
#> 21                   46220              NA                     NA
#> 22                   20020              NA                    222
#> 23                   20020              NA                    222
#> 24                   42460              NA                    174
#> 25                   13820              NA                    142
#> 26                   22520              NA                    250
#> 27                   19460              NA                    290
#> 28                   12220              NA                    194
#> 29                   26620              NA                    290
#> 30                   33860              NA                    388
#> 31                   12220              NA                    194
#> 32                   26620              NA                    290
#> 33                   10700              NA                    290
#> 34                   33660              NA                    380
#> 35                   33860              NA                    388
#> 36                   19460              NA                    290
#> 37                   46220              NA                     NA
#> 38                   45980              NA                     NA
#> 39                   17980              NA                    194
#> 40                   13820              NA                    142
#> 41                   13820              NA                    142
#> 42                   45180              NA                    142
#> 43                   10760              NA                    194
#> 44                   46220              NA                     NA
#> 45                   13820              NA                    142
#> 46                   10580              NA                    104
#> 47                   35620           35614                    408
#> 48                   13780              NA                     NA
#> 49                   36460              NA                    160
#> 50                   12180              NA                    532
#> 51                   27460              NA                     NA
#> 52                   21300              NA                    236
#> 53                   38460              NA                     NA
#> 54                   26460              NA                    104
#> 55                   18660              NA                    296
#> 56                   28880              NA                    408
#> 57                   15380              NA                    160
#> 58                   24100              NA                    104
#> 59                   12860              NA                    464
#> 60                   46540              NA                     NA
#> 61                   48060              NA                     NA
#> 62                   35620           35614                    408
#> 63                   40380              NA                    464
#> 64                   45060              NA                    532
#> 65                   40380              NA                    464
#> 66                   11220              NA                    104
#> 67                   35620           35004                    408
#> 68                   35620           35614                    408
#> 69                   15380              NA                    160
#> 70                   46540              NA                     NA
#> 71                   45060              NA                    532
#> 72                   40380              NA                    464
#> 73                   28880              NA                    408
#> 74                   40380              NA                    464
#> 75                   45060              NA                    532
#> 76                   36580              NA                     NA
#> 77                   35620           35614                    408
#> 78                   35620           35614                    408
#> 79                   10580              NA                    104
#> 80                   35620           35614                    408
#> 81                   35620           35614                    408
#> 82                   32390              NA                     NA
#> 83                   10580              NA                    104
#> 84                   10580              NA                    104
#> 85                   10580              NA                    104
#> 86                   42900              NA                    464
#> 87                   18500              NA                    236
#> 88                   35620           35004                    408
#> 89                   33910              NA                    408
#> 90                   13780              NA                     NA
#> 91                   27060              NA                    296
#> 92                   28740              NA                    408
#> 93                   24020              NA                    104
#> 94                   24020              NA                    104
#> 95                   40380              NA                    464
#> 96                   35620           35614                    408
#>                   metro.census.cbsa.name              metro.micro.name
#> 1                         Montgomery, AL Metropolitan Statistical Area
#> 2              Daphne-Fairhope-Foley, AL Metropolitan Statistical Area
#> 3                         Eufaula, AL-GA Micropolitan Statistical Area
#> 4                         Birmingham, AL Metropolitan Statistical Area
#> 5                         Birmingham, AL Metropolitan Statistical Area
#> 6                    Anniston-Oxford, AL Metropolitan Statistical Area
#> 7                        LaGrange, GA-AL Micropolitan Statistical Area
#> 8                         Birmingham, AL Metropolitan Statistical Area
#> 9                         Enterprise, AL Micropolitan Statistical Area
#> 10            Florence-Muscle Shoals, AL Metropolitan Statistical Area
#> 11               Talladega-Sylacauga, AL Micropolitan Statistical Area
#> 12                           Cullman, AL Micropolitan Statistical Area
#> 13                             Ozark, AL Micropolitan Statistical Area
#> 14                             Selma, AL Micropolitan Statistical Area
#> 15                        Fort Payne, AL Micropolitan Statistical Area
#> 16                        Montgomery, AL Metropolitan Statistical Area
#> 17                           Gadsden, AL Metropolitan Statistical Area
#> 18                      Russellville, AL Micropolitan Statistical Area
#> 19                            Dothan, AL Metropolitan Statistical Area
#> 20                        Tuscaloosa, AL Metropolitan Statistical Area
#> 21                        Tuscaloosa, AL Metropolitan Statistical Area
#> 22                            Dothan, AL Metropolitan Statistical Area
#> 23                            Dothan, AL Metropolitan Statistical Area
#> 24                        Scottsboro, AL Micropolitan Statistical Area
#> 25                        Birmingham, AL Metropolitan Statistical Area
#> 26            Florence-Muscle Shoals, AL Metropolitan Statistical Area
#> 27                           Decatur, AL Metropolitan Statistical Area
#> 28                    Auburn-Opelika, AL Metropolitan Statistical Area
#> 29                        Huntsville, AL Metropolitan Statistical Area
#> 30                        Montgomery, AL Metropolitan Statistical Area
#> 31                    Auburn-Opelika, AL Metropolitan Statistical Area
#> 32                        Huntsville, AL Metropolitan Statistical Area
#> 33                       Albertville, AL Micropolitan Statistical Area
#> 34                            Mobile, AL Metropolitan Statistical Area
#> 35                        Montgomery, AL Metropolitan Statistical Area
#> 36                           Decatur, AL Metropolitan Statistical Area
#> 37                        Tuscaloosa, AL Metropolitan Statistical Area
#> 38                              Troy, AL Micropolitan Statistical Area
#> 39                       Columbus, GA-AL Metropolitan Statistical Area
#> 40                        Birmingham, AL Metropolitan Statistical Area
#> 41                        Birmingham, AL Metropolitan Statistical Area
#> 42               Talladega-Sylacauga, AL Micropolitan Statistical Area
#> 43                    Alexander City, AL Micropolitan Statistical Area
#> 44                        Tuscaloosa, AL Metropolitan Statistical Area
#> 45                        Birmingham, AL Metropolitan Statistical Area
#> 46           Albany-Schenectady-Troy, NY Metropolitan Statistical Area
#> 47    New York-Newark-Jersey City, NY-NJ Metropolitan Statistical Area
#> 48                        Binghamton, NY Metropolitan Statistical Area
#> 49                             Olean, NY Micropolitan Statistical Area
#> 50                            Auburn, NY Micropolitan Statistical Area
#> 51                 Jamestown-Dunkirk, NY Micropolitan Statistical Area
#> 52                            Elmira, NY Metropolitan Statistical Area
#> 53                       Plattsburgh, NY Micropolitan Statistical Area
#> 54                            Hudson, NY Micropolitan Statistical Area
#> 55                          Cortland, NY Micropolitan Statistical Area
#> 56 Kiryas Joel-Poughkeepsie-Newburgh, NY Metropolitan Statistical Area
#> 57               Buffalo-Cheektowaga, NY Metropolitan Statistical Area
#> 58                      Gloversville, NY Micropolitan Statistical Area
#> 59                           Batavia, NY Micropolitan Statistical Area
#> 60                        Utica-Rome, NY Metropolitan Statistical Area
#> 61               Watertown-Fort Drum, NY Metropolitan Statistical Area
#> 62    New York-Newark-Jersey City, NY-NJ Metropolitan Statistical Area
#> 63                         Rochester, NY Metropolitan Statistical Area
#> 64                          Syracuse, NY Metropolitan Statistical Area
#> 65                         Rochester, NY Metropolitan Statistical Area
#> 66                         Amsterdam, NY Micropolitan Statistical Area
#> 67    New York-Newark-Jersey City, NY-NJ Metropolitan Statistical Area
#> 68    New York-Newark-Jersey City, NY-NJ Metropolitan Statistical Area
#> 69               Buffalo-Cheektowaga, NY Metropolitan Statistical Area
#> 70                        Utica-Rome, NY Metropolitan Statistical Area
#> 71                          Syracuse, NY Metropolitan Statistical Area
#> 72                         Rochester, NY Metropolitan Statistical Area
#> 73 Kiryas Joel-Poughkeepsie-Newburgh, NY Metropolitan Statistical Area
#> 74                         Rochester, NY Metropolitan Statistical Area
#> 75                          Syracuse, NY Metropolitan Statistical Area
#> 76                           Oneonta, NY Micropolitan Statistical Area
#> 77    New York-Newark-Jersey City, NY-NJ Metropolitan Statistical Area
#> 78    New York-Newark-Jersey City, NY-NJ Metropolitan Statistical Area
#> 79           Albany-Schenectady-Troy, NY Metropolitan Statistical Area
#> 80    New York-Newark-Jersey City, NY-NJ Metropolitan Statistical Area
#> 81    New York-Newark-Jersey City, NY-NJ Metropolitan Statistical Area
#> 82                Massena-Ogdensburg, NY Micropolitan Statistical Area
#> 83           Albany-Schenectady-Troy, NY Metropolitan Statistical Area
#> 84           Albany-Schenectady-Troy, NY Metropolitan Statistical Area
#> 85           Albany-Schenectady-Troy, NY Metropolitan Statistical Area
#> 86                      Seneca Falls, NY Micropolitan Statistical Area
#> 87                           Corning, NY Micropolitan Statistical Area
#> 88    New York-Newark-Jersey City, NY-NJ Metropolitan Statistical Area
#> 89                        Monticello, NY Micropolitan Statistical Area
#> 90                        Binghamton, NY Metropolitan Statistical Area
#> 91                            Ithaca, NY Metropolitan Statistical Area
#> 92                          Kingston, NY Metropolitan Statistical Area
#> 93                       Glens Falls, NY Metropolitan Statistical Area
#> 94                       Glens Falls, NY Metropolitan Statistical Area
#> 95                         Rochester, NY Metropolitan Statistical Area
#> 96    New York-Newark-Jersey City, NY-NJ Metropolitan Statistical Area
#>                              metro.div.name
#> 1                                          
#> 2                                          
#> 3                                          
#> 4                                          
#> 5                                          
#> 6                                          
#> 7                                          
#> 8                                          
#> 9                                          
#> 10                                         
#> 11                                         
#> 12                                         
#> 13                                         
#> 14                                         
#> 15                                         
#> 16                                         
#> 17                                         
#> 18                                         
#> 19                                         
#> 20                                         
#> 21                                         
#> 22                                         
#> 23                                         
#> 24                                         
#> 25                                         
#> 26                                         
#> 27                                         
#> 28                                         
#> 29                                         
#> 30                                         
#> 31                                         
#> 32                                         
#> 33                                         
#> 34                                         
#> 35                                         
#> 36                                         
#> 37                                         
#> 38                                         
#> 39                                         
#> 40                                         
#> 41                                         
#> 42                                         
#> 43                                         
#> 44                                         
#> 45                                         
#> 46                                         
#> 47 New York-Jersey City-White Plains, NY-NJ
#> 48                                         
#> 49                                         
#> 50                                         
#> 51                                         
#> 52                                         
#> 53                                         
#> 54                                         
#> 55                                         
#> 56                                         
#> 57                                         
#> 58                                         
#> 59                                         
#> 60                                         
#> 61                                         
#> 62 New York-Jersey City-White Plains, NY-NJ
#> 63                                         
#> 64                                         
#> 65                                         
#> 66                                         
#> 67         Nassau County-Suffolk County, NY
#> 68 New York-Jersey City-White Plains, NY-NJ
#> 69                                         
#> 70                                         
#> 71                                         
#> 72                                         
#> 73                                         
#> 74                                         
#> 75                                         
#> 76                                         
#> 77 New York-Jersey City-White Plains, NY-NJ
#> 78 New York-Jersey City-White Plains, NY-NJ
#> 79                                         
#> 80 New York-Jersey City-White Plains, NY-NJ
#> 81 New York-Jersey City-White Plains, NY-NJ
#> 82                                         
#> 83                                         
#> 84                                         
#> 85                                         
#> 86                                         
#> 87                                         
#> 88         Nassau County-Suffolk County, NY
#> 89                                         
#> 90                                         
#> 91                                         
#> 92                                         
#> 93                                         
#> 94                                         
#> 95                                         
#> 96 New York-Jersey City-White Plains, NY-NJ
#>                                  metro.census.csa.name  census.county.name
#> 1                                 Montgomery-Selma, AL      Autauga County
#> 2                           Mobile-Daphne-Fairhope, AL      Baldwin County
#> 3                                                           Barbour County
#> 4                     Birmingham-Cullman-Talladega, AL         Bibb County
#> 5                     Birmingham-Cullman-Talladega, AL       Blount County
#> 6                                                           Calhoun County
#> 7  Atlanta--Athens-Clarke County--Sandy Springs, GA-AL     Chambers County
#> 8                     Birmingham-Cullman-Talladega, AL      Chilton County
#> 9                          Dothan-Enterprise-Ozark, AL       Coffee County
#> 10             Florence-Muscle Shoals-Russellville, AL      Colbert County
#> 11                    Birmingham-Cullman-Talladega, AL        Coosa County
#> 12                    Birmingham-Cullman-Talladega, AL      Cullman County
#> 13                         Dothan-Enterprise-Ozark, AL         Dale County
#> 14                                Montgomery-Selma, AL       Dallas County
#> 15               Huntsville-Decatur-Albertville, AL-TN       DeKalb County
#> 16                                Montgomery-Selma, AL       Elmore County
#> 17                                                           Etowah County
#> 18             Florence-Muscle Shoals-Russellville, AL     Franklin County
#> 19                         Dothan-Enterprise-Ozark, AL       Geneva County
#> 20                                                           Greene County
#> 21                                                             Hale County
#> 22                         Dothan-Enterprise-Ozark, AL        Henry County
#> 23                         Dothan-Enterprise-Ozark, AL      Houston County
#> 24              Chattanooga-Cleveland-Dalton, TN-GA-AL      Jackson County
#> 25                    Birmingham-Cullman-Talladega, AL    Jefferson County
#> 26             Florence-Muscle Shoals-Russellville, AL   Lauderdale County
#> 27               Huntsville-Decatur-Albertville, AL-TN     Lawrence County
#> 28                      Columbus-Auburn-Opelika, GA-AL          Lee County
#> 29               Huntsville-Decatur-Albertville, AL-TN    Limestone County
#> 30                                Montgomery-Selma, AL      Lowndes County
#> 31                      Columbus-Auburn-Opelika, GA-AL        Macon County
#> 32               Huntsville-Decatur-Albertville, AL-TN      Madison County
#> 33               Huntsville-Decatur-Albertville, AL-TN     Marshall County
#> 34                          Mobile-Daphne-Fairhope, AL       Mobile County
#> 35                                Montgomery-Selma, AL   Montgomery County
#> 36               Huntsville-Decatur-Albertville, AL-TN       Morgan County
#> 37                                                          Pickens County
#> 38                                                             Pike County
#> 39                      Columbus-Auburn-Opelika, GA-AL      Russell County
#> 40                    Birmingham-Cullman-Talladega, AL    St. Clair County
#> 41                    Birmingham-Cullman-Talladega, AL       Shelby County
#> 42                    Birmingham-Cullman-Talladega, AL    Talladega County
#> 43                      Columbus-Auburn-Opelika, GA-AL   Tallapoosa County
#> 44                                                       Tuscaloosa County
#> 45                    Birmingham-Cullman-Talladega, AL       Walker County
#> 46                              Albany-Schenectady, NY       Albany County
#> 47                        New York-Newark, NY-NJ-CT-PA        Bronx County
#> 48                                                           Broome County
#> 49                       Buffalo-Cheektowaga-Olean, NY  Cattaraugus County
#> 50                                 Syracuse-Auburn, NY       Cayuga County
#> 51                                                       Chautauqua County
#> 52                                  Elmira-Corning, NY      Chemung County
#> 53                                                          Clinton County
#> 54                              Albany-Schenectady, NY     Columbia County
#> 55                                 Ithaca-Cortland, NY     Cortland County
#> 56                        New York-Newark, NY-NJ-CT-PA     Dutchess County
#> 57                       Buffalo-Cheektowaga-Olean, NY         Erie County
#> 58                              Albany-Schenectady, NY       Fulton County
#> 59                  Rochester-Batavia-Seneca Falls, NY      Genesee County
#> 60                                                         Herkimer County
#> 61                                                        Jefferson County
#> 62                        New York-Newark, NY-NJ-CT-PA        Kings County
#> 63                  Rochester-Batavia-Seneca Falls, NY   Livingston County
#> 64                                 Syracuse-Auburn, NY      Madison County
#> 65                  Rochester-Batavia-Seneca Falls, NY       Monroe County
#> 66                              Albany-Schenectady, NY   Montgomery County
#> 67                        New York-Newark, NY-NJ-CT-PA       Nassau County
#> 68                        New York-Newark, NY-NJ-CT-PA     New York County
#> 69                       Buffalo-Cheektowaga-Olean, NY      Niagara County
#> 70                                                           Oneida County
#> 71                                 Syracuse-Auburn, NY     Onondaga County
#> 72                  Rochester-Batavia-Seneca Falls, NY      Ontario County
#> 73                        New York-Newark, NY-NJ-CT-PA       Orange County
#> 74                  Rochester-Batavia-Seneca Falls, NY      Orleans County
#> 75                                 Syracuse-Auburn, NY       Oswego County
#> 76                                                           Otsego County
#> 77                        New York-Newark, NY-NJ-CT-PA       Putnam County
#> 78                        New York-Newark, NY-NJ-CT-PA       Queens County
#> 79                              Albany-Schenectady, NY   Rensselaer County
#> 80                        New York-Newark, NY-NJ-CT-PA     Richmond County
#> 81                        New York-Newark, NY-NJ-CT-PA     Rockland County
#> 82                                                     St. Lawrence County
#> 83                              Albany-Schenectady, NY     Saratoga County
#> 84                              Albany-Schenectady, NY  Schenectady County
#> 85                              Albany-Schenectady, NY    Schoharie County
#> 86                  Rochester-Batavia-Seneca Falls, NY       Seneca County
#> 87                                  Elmira-Corning, NY      Steuben County
#> 88                        New York-Newark, NY-NJ-CT-PA      Suffolk County
#> 89                        New York-Newark, NY-NJ-CT-PA     Sullivan County
#> 90                                                            Tioga County
#> 91                                 Ithaca-Cortland, NY     Tompkins County
#> 92                        New York-Newark, NY-NJ-CT-PA       Ulster County
#> 93                              Albany-Schenectady, NY       Warren County
#> 94                              Albany-Schenectady, NY   Washington County
#> 95                  Rochester-Batavia-Seneca Falls, NY        Wayne County
#> 96                        New York-Newark, NY-NJ-CT-PA  Westchester County
#>    state.census.name state.censu.geoid census.county.geoid census.centrout.name
#> 1            Alabama                 1                   1              Central
#> 2            Alabama                 1                   3              Central
#> 3            Alabama                 1                   5              Central
#> 4            Alabama                 1                   7             Outlying
#> 5            Alabama                 1                   9             Outlying
#> 6            Alabama                 1                  15              Central
#> 7            Alabama                 1                  17             Outlying
#> 8            Alabama                 1                  21             Outlying
#> 9            Alabama                 1                  31              Central
#> 10           Alabama                 1                  33              Central
#> 11           Alabama                 1                  37             Outlying
#> 12           Alabama                 1                  43              Central
#> 13           Alabama                 1                  45              Central
#> 14           Alabama                 1                  47              Central
#> 15           Alabama                 1                  49              Central
#> 16           Alabama                 1                  51              Central
#> 17           Alabama                 1                  55              Central
#> 18           Alabama                 1                  59              Central
#> 19           Alabama                 1                  61             Outlying
#> 20           Alabama                 1                  63             Outlying
#> 21           Alabama                 1                  65             Outlying
#> 22           Alabama                 1                  67             Outlying
#> 23           Alabama                 1                  69              Central
#> 24           Alabama                 1                  71              Central
#> 25           Alabama                 1                  73              Central
#> 26           Alabama                 1                  77              Central
#> 27           Alabama                 1                  79             Outlying
#> 28           Alabama                 1                  81              Central
#> 29           Alabama                 1                  83              Central
#> 30           Alabama                 1                  85             Outlying
#> 31           Alabama                 1                  87             Outlying
#> 32           Alabama                 1                  89              Central
#> 33           Alabama                 1                  95              Central
#> 34           Alabama                 1                  97              Central
#> 35           Alabama                 1                 101              Central
#> 36           Alabama                 1                 103              Central
#> 37           Alabama                 1                 107             Outlying
#> 38           Alabama                 1                 109              Central
#> 39           Alabama                 1                 113              Central
#> 40           Alabama                 1                 115              Central
#> 41           Alabama                 1                 117              Central
#> 42           Alabama                 1                 121              Central
#> 43           Alabama                 1                 123              Central
#> 44           Alabama                 1                 125              Central
#> 45           Alabama                 1                 127             Outlying
#> 46          New York                36                   1              Central
#> 47          New York                36                   5              Central
#> 48          New York                36                   7              Central
#> 49          New York                36                   9              Central
#> 50          New York                36                  11              Central
#> 51          New York                36                  13              Central
#> 52          New York                36                  15              Central
#> 53          New York                36                  19              Central
#> 54          New York                36                  21              Central
#> 55          New York                36                  23              Central
#> 56          New York                36                  27              Central
#> 57          New York                36                  29              Central
#> 58          New York                36                  35              Central
#> 59          New York                36                  37              Central
#> 60          New York                36                  43             Outlying
#> 61          New York                36                  45              Central
#> 62          New York                36                  47              Central
#> 63          New York                36                  51             Outlying
#> 64          New York                36                  53             Outlying
#> 65          New York                36                  55              Central
#> 66          New York                36                  57              Central
#> 67          New York                36                  59              Central
#> 68          New York                36                  61              Central
#> 69          New York                36                  63              Central
#> 70          New York                36                  65              Central
#> 71          New York                36                  67              Central
#> 72          New York                36                  69              Central
#> 73          New York                36                  71              Central
#> 74          New York                36                  73             Outlying
#> 75          New York                36                  75             Outlying
#> 76          New York                36                  77              Central
#> 77          New York                36                  79              Central
#> 78          New York                36                  81              Central
#> 79          New York                36                  83              Central
#> 80          New York                36                  85              Central
#> 81          New York                36                  87              Central
#> 82          New York                36                  89              Central
#> 83          New York                36                  91             Outlying
#> 84          New York                36                  93              Central
#> 85          New York                36                  95             Outlying
#> 86          New York                36                  99              Central
#> 87          New York                36                 101              Central
#> 88          New York                36                 103              Central
#> 89          New York                36                 105              Central
#> 90          New York                36                 107              Central
#> 91          New York                36                 109              Central
#> 92          New York                36                 111              Central
#> 93          New York                36                 113              Central
#> 94          New York                36                 115              Central
#> 95          New York                36                 117             Outlying
#> 96          New York                36                 119              Central
#>    state.census.abbr
#> 1                 AL
#> 2                 AL
#> 3                 AL
#> 4                 AL
#> 5                 AL
#> 6                 AL
#> 7                 AL
#> 8                 AL
#> 9                 AL
#> 10                AL
#> 11                AL
#> 12                AL
#> 13                AL
#> 14                AL
#> 15                AL
#> 16                AL
#> 17                AL
#> 18                AL
#> 19                AL
#> 20                AL
#> 21                AL
#> 22                AL
#> 23                AL
#> 24                AL
#> 25                AL
#> 26                AL
#> 27                AL
#> 28                AL
#> 29                AL
#> 30                AL
#> 31                AL
#> 32                AL
#> 33                AL
#> 34                AL
#> 35                AL
#> 36                AL
#> 37                AL
#> 38                AL
#> 39                AL
#> 40                AL
#> 41                AL
#> 42                AL
#> 43                AL
#> 44                AL
#> 45                AL
#> 46                NY
#> 47                NY
#> 48                NY
#> 49                NY
#> 50                NY
#> 51                NY
#> 52                NY
#> 53                NY
#> 54                NY
#> 55                NY
#> 56                NY
#> 57                NY
#> 58                NY
#> 59                NY
#> 60                NY
#> 61                NY
#> 62                NY
#> 63                NY
#> 64                NY
#> 65                NY
#> 66                NY
#> 67                NY
#> 68                NY
#> 69                NY
#> 70                NY
#> 71                NY
#> 72                NY
#> 73                NY
#> 74                NY
#> 75                NY
#> 76                NY
#> 77                NY
#> 78                NY
#> 79                NY
#> 80                NY
#> 81                NY
#> 82                NY
#> 83                NY
#> 84                NY
#> 85                NY
#> 86                NY
#> 87                NY
#> 88                NY
#> 89                NY
#> 90                NY
#> 91                NY
#> 92                NY
#> 93                NY
#> 94                NY
#> 95                NY
#> 96                NY
#>