Provide documentation for NTEE2 codes

Take user inputs across levels 1- 5 of the NTEE2 code structure and returns text metadata describing each NTEE Code level

Usage

ntee_preview(ntee, ntee.group, ntee.code, ntee.orgtype)

Arguments

ntee

character vector. Vector of user inputs. The user inputs are progressively filtered until group, code and orgtypes are sorted into separate vectors.

ntee.group

character vector. Vector of desired Industry Group codes to filter. Use "all" to include all possible codes.

ntee.code

character vector. Sequence of desired Industry, Division and Subdivision codes (old code structure) to use in filtering. Use "all" to include all possible codes. Can also provide only partial codes. For example "A" or "Axx" will query NTEE2 codes based on Industry group "A" and all division and subdivisions.

ntee.orgtype

character vector. Vector of Organization Types. Use "all" to include all possible codes.

Value

text strings that describe NTEE2 codes at each level

Examples

ntee_preview(ntee = c("ART", "A2X"))
#> Collecting Matching Industry Groups
#> Collecting Matching Industry Division and Subdivisions
#> Collecting Matching Organization Types
#> Collecting Matching Industry Groups
#> Collecting Matching Industry Division and Subdivisions
#> Collecting Matching Organization Types
#> 
#> 
#>  ART: Arts, Culture & Humanities
#> 
#>      AA: Alliance/Advocacy Organizations
#> 
#>          A00: 
#> 
#>          
#> 
#>          A61: 
#> 
#>          Providing support to organizations that operate facilities including
#>          theaters for the performing arts.
#> 
#>      MM: Monetary Support - Multiple Organizations
#> 
#>          A00: 
#> 
#>          
#> 
#>          A26: 
#> 
#>          Providing support to organizations that foster, nurture and sustain
#>          artistic excellence and create a climate in which the arts and
#>          humanities may flourish in a community.
#> 
#>          A30: 
#> 
#>          Providing support to organizations that are involved in publishing,
#>          productions or programming utilizing any of a variety of media or in
#>          educational programs related to the media. Use this code for
#>          organizations involved in productions in different types of media or
#>          for those using types of media not specified below.
#> 
#>          A32: 
#> 
#>          Providing support to organizations engaged in television productions
#>          and broadcast. Includes producers of public television, cable
#>          television, community access television, broadcasting in schools, and
#>          interactive television.
#> 
#>          A60: 
#> 
#>          Providing support to organizations that provide access to the
#>          performing arts or a variety of art forms including the performing
#>          arts, and/or which present performing arts series. Use this code for
#>          organizations devoted to art forms not specified below.
#> 
#>          A61: 
#> 
#>          Providing support to organizations that operate facilities including
#>          theaters for the performing arts.
#> 
#>          A65: 
#> 
#>          Providing support to organizations whose primary activity is the
#>          production of plays. (Organizations that present the productions of
#>          others should be classified as presenters. (A61)
#> 
#>      MS: Monetary Support - Single Organization
#> 
#>          A00: 
#> 
#>          
#> 
#>          A20: 
#> 
#>          Providing support to organizations that promote, produce or provide
#>          access to a variety of arts experiences encompassing the visual, media
#>          or performing arts.
#> 
#>          A32: 
#> 
#>          Providing support to organizations engaged in television productions
#>          and broadcast. Includes producers of public television, cable
#>          television, community access television, broadcasting in schools, and
#>          interactive television.
#> 
#>          A40: 
#> 
#>          Providing support to organizations whose activities promote or provide
#>          access to a variety of visual arts. Included are organizations that
#>          focus on an art form in a specific medium such as stained glass or
#>          textiles.
#> 
#>          A50: 
#> 
#>          Providing support to organizations that acquire, preserve, research,
#>          exhibit and provide for the educational use of works of art, objects or
#>          artifacts that are related to the study of zoology, biology, botany,
#>          mineralogy, geology and other natural sciences; history; archeology; or
#>          science and technology. Use this code for types of museums not
#>          specified below.
#> 
#>          A51: 
#> 
#>          Providing support to organizations that acquire, preserve and exhibit
#>          collections of objects including fine art, decorative art, folk/ethnic
#>          art and textiles that are collected primarily for their aesthetic
#>          qualities and their importance as representatives of a particular
#>          artistic tradition or style.
#> 
#>          A52: 
#> 
#>          Providing support to organizations that develop and maintain
#>          educational exhibits that are structured for the size, interests and
#>          intellectual capacity of young children. Childrens museums are
#>          participatory and allow visitors to interact with the exhibits
#>          (touching, exploring, manipulating), and many offer organized lessons
#>          for groups of school children that focus on one or more of the
#>          exhibits.  NOTE: See also A57, Science and Technology Museums, for
#>          other organizations with substantial activities and collections for
#>          children.
#> 
#>          A54: 
#> 
#>          Providing support to organizations that acquire, preserve, research and
#>          exhibit collections of objects including documents, tools, implements
#>          and furnishings that have significance in helping to interpret or
#>          understand the past. History museums may specialize in a specific era
#>          such as early Greece or Rome, a particular geographical region such as
#>          California or Appalachia, a particular ethnic or cultural group such as
#>          Native Americans or a specific subject area such as costumes; and may
#>          contain items created or used by contemporary or historical figures.
#> 
#>          A60: 
#> 
#>          Providing support to organizations that provide access to the
#>          performing arts or a variety of art forms including the performing
#>          arts, and/or which present performing arts series. Use this code for
#>          organizations devoted to art forms not specified below.
#> 
#>          A61: 
#> 
#>          Providing support to organizations that operate facilities including
#>          theaters for the performing arts.
#> 
#>          A62: 
#> 
#>          Providing support to organizations that create and/or perform a variety
#>          of types of dance, including modern, folk, country and western, and
#>          square dancing.
#> 
#>          A65: 
#> 
#>          Providing support to organizations whose primary activity is the
#>          production of plays. (Organizations that present the productions of
#>          others should be classified as presenters. (A61)
#> 
#>          A69: 
#> 
#>          Providing support to organizations whose primary activity is the
#>          operation of a symphony
#> 
#>          A70: 
#> 
#>          Providing support to organizations that promote the study, teaching, or
#>          appreciation of the humanities (archaeology, art history, history,
#>          modern and classical languages, philosophy and ethics, and theology and
#>          comparative religion).
#> 
#>          A80: 
#> 
#>          Providing support to organizations that promote awareness of and
#>          appreciation for history and historical artifacts.  Use this code for
#>          organizations that are not specified below.
#> 
#>      MT: Management and Technical Assistance
#> 
#>          A00: 
#> 
#>          
#> 
#>          A26: 
#> 
#>          Providing support to organizations that foster, nurture and sustain
#>          artistic excellence and create a climate in which the arts and
#>          humanities may flourish in a community.
#> 
#>          A54: 
#> 
#>          Providing support to organizations that acquire, preserve, research and
#>          exhibit collections of objects including documents, tools, implements
#>          and furnishings that have significance in helping to interpret or
#>          understand the past. History museums may specialize in a specific era
#>          such as early Greece or Rome, a particular geographical region such as
#>          California or Appalachia, a particular ethnic or cultural group such as
#>          Native Americans or a specific subject area such as costumes; and may
#>          contain items created or used by contemporary or historical figures.
#> 
#>          A61: 
#> 
#>          Providing support to organizations that operate facilities including
#>          theaters for the performing arts.
#> 
#>          A82: 
#> 
#>          Providing support to organizations that interpret, collect and preserve
#>          the historical heritage of a specific geographic location.
#> 
#>          A90: 
#> 
#>          Providing support to organizations that provide services such as
#>          facilities or touring assistance for arts organizations.
#> 
#>      NS: Nonmonetary Support Not Elsewhere Classified (N.E.C.)
#> 
#>          A00: 
#> 
#>          
#> 
#>          A20: 
#> 
#>          Providing support to organizations that promote, produce or provide
#>          access to a variety of arts experiences encompassing the visual, media
#>          or performing arts.
#> 
#>          A61: 
#> 
#>          Providing support to organizations that operate facilities including
#>          theaters for the performing arts.
#> 
#>          A69: 
#> 
#>          Providing support to organizations whose primary activity is the
#>          operation of a symphony
#> 
#>      PA: Professional Societies/Associations
#> 
#>          A00: 
#> 
#>          
#> 
#>          A20: 
#> 
#>          Providing support to organizations that promote, produce or provide
#>          access to a variety of arts experiences encompassing the visual, media
#>          or performing arts.
#> 
#>          A25: 
#> 
#>          Providing support to organizations that provide informal arts
#>          educational programming and/or instruction but do not grant diplomas or
#>          degrees; or which offer services regarding the arts to educational
#>          institutions or to public entities involved in education.
#> 
#>          A30: 
#> 
#>          Providing support to organizations that are involved in publishing,
#>          productions or programming utilizing any of a variety of media or in
#>          educational programs related to the media. Use this code for
#>          organizations involved in productions in different types of media or
#>          for those using types of media not specified below.
#> 
#>          A33: 
#> 
#>          Providing support to organizations that publish a variety of materials
#>          on diverse topics. Includes university presses.
#> 
#>          A40: 
#> 
#>          Providing support to organizations whose activities promote or provide
#>          access to a variety of visual arts. Included are organizations that
#>          focus on an art form in a specific medium such as stained glass or
#>          textiles.
#> 
#>          A50: 
#> 
#>          Providing support to organizations that acquire, preserve, research,
#>          exhibit and provide for the educational use of works of art, objects or
#>          artifacts that are related to the study of zoology, biology, botany,
#>          mineralogy, geology and other natural sciences; history; archeology; or
#>          science and technology. Use this code for types of museums not
#>          specified below.
#> 
#>          A60: 
#> 
#>          Providing support to organizations that provide access to the
#>          performing arts or a variety of art forms including the performing
#>          arts, and/or which present performing arts series. Use this code for
#>          organizations devoted to art forms not specified below.
#> 
#>          A61: 
#> 
#>          Providing support to organizations that operate facilities including
#>          theaters for the performing arts.
#> 
#>          A65: 
#> 
#>          Providing support to organizations whose primary activity is the
#>          production of plays. (Organizations that present the productions of
#>          others should be classified as presenters. (A61)
#> 
#>          A68: 
#> 
#>          Providing support to organizations whose primary activity is the
#>          production of musical events, including concert series and music
#>          festivals.
#> 
#>          A69: 
#> 
#>          Providing support to organizations whose primary activity is the
#>          operation of a symphony
#> 
#>          A70: 
#> 
#>          Providing support to organizations that promote the study, teaching, or
#>          appreciation of the humanities (archaeology, art history, history,
#>          modern and classical languages, philosophy and ethics, and theology and
#>          comparative religion).
#> 
#>          A80: 
#> 
#>          Providing support to organizations that promote awareness of and
#>          appreciation for history and historical artifacts.  Use this code for
#>          organizations that are not specified below.
#> 
#>      RG: Organizations that promote, produce or provide access to a variety of
#>          arts experiences encompassing the visual, media or performing arts.
#> 
#>          A00: 
#> 
#>          
#> 
#>          A20: 
#> 
#>          Providing support to organizations that promote, produce or provide
#>          access to a variety of arts experiences encompassing the visual, media
#>          or performing arts.
#> 
#>          A23: 
#> 
#>          
#> 
#>          A24: 
#> 
#>          
#> 
#>          A25: 
#> 
#>          Providing support to organizations that provide informal arts
#>          educational programming and/or instruction but do not grant diplomas or
#>          degrees; or which offer services regarding the arts to educational
#>          institutions or to public entities involved in education.
#> 
#>          A26: 
#> 
#>          Providing support to organizations that foster, nurture and sustain
#>          artistic excellence and create a climate in which the arts and
#>          humanities may flourish in a community.
#> 
#>          A27: 
#> 
#>          
#> 
#>          A30: 
#> 
#>          Providing support to organizations that are involved in publishing,
#>          productions or programming utilizing any of a variety of media or in
#>          educational programs related to the media. Use this code for
#>          organizations involved in productions in different types of media or
#>          for those using types of media not specified below.
#> 
#>          A31: 
#> 
#>          
#> 
#>          A32: 
#> 
#>          Providing support to organizations engaged in television productions
#>          and broadcast. Includes producers of public television, cable
#>          television, community access television, broadcasting in schools, and
#>          interactive television.
#> 
#>          A33: 
#> 
#>          Providing support to organizations that publish a variety of materials
#>          on diverse topics. Includes university presses.
#> 
#>          A34: 
#> 
#>          Providing support to organizations engaged in radio productions and
#>          broadcast, including public radio stations and radio stations that
#>          offer programming that is designed for community residents who are
#>          blind or visually impaired (radio reading services).
#> 
#>          A40: 
#> 
#>          Providing support to organizations whose activities promote or provide
#>          access to a variety of visual arts. Included are organizations that
#>          focus on an art form in a specific medium such as stained glass or
#>          textiles.
#> 
#>          A50: 
#> 
#>          Providing support to organizations that acquire, preserve, research,
#>          exhibit and provide for the educational use of works of art, objects or
#>          artifacts that are related to the study of zoology, biology, botany,
#>          mineralogy, geology and other natural sciences; history; archeology; or
#>          science and technology. Use this code for types of museums not
#>          specified below.
#> 
#>          A51: 
#> 
#>          Providing support to organizations that acquire, preserve and exhibit
#>          collections of objects including fine art, decorative art, folk/ethnic
#>          art and textiles that are collected primarily for their aesthetic
#>          qualities and their importance as representatives of a particular
#>          artistic tradition or style.
#> 
#>          A52: 
#> 
#>          Providing support to organizations that develop and maintain
#>          educational exhibits that are structured for the size, interests and
#>          intellectual capacity of young children. Childrens museums are
#>          participatory and allow visitors to interact with the exhibits
#>          (touching, exploring, manipulating), and many offer organized lessons
#>          for groups of school children that focus on one or more of the
#>          exhibits.  NOTE: See also A57, Science and Technology Museums, for
#>          other organizations with substantial activities and collections for
#>          children.
#> 
#>          A53: 
#> 
#>          
#> 
#>          A54: 
#> 
#>          Providing support to organizations that acquire, preserve, research and
#>          exhibit collections of objects including documents, tools, implements
#>          and furnishings that have significance in helping to interpret or
#>          understand the past. History museums may specialize in a specific era
#>          such as early Greece or Rome, a particular geographical region such as
#>          California or Appalachia, a particular ethnic or cultural group such as
#>          Native Americans or a specific subject area such as costumes; and may
#>          contain items created or used by contemporary or historical figures.
#> 
#>          A56: 
#> 
#>          
#> 
#>          A57: 
#> 
#>          
#> 
#>          A60: 
#> 
#>          Providing support to organizations that provide access to the
#>          performing arts or a variety of art forms including the performing
#>          arts, and/or which present performing arts series. Use this code for
#>          organizations devoted to art forms not specified below.
#> 
#>          A61: 
#> 
#>          Providing support to organizations that operate facilities including
#>          theaters for the performing arts.
#> 
#>          A62: 
#> 
#>          Providing support to organizations that create and/or perform a variety
#>          of types of dance, including modern, folk, country and western, and
#>          square dancing.
#> 
#>          A63: 
#> 
#>          
#> 
#>          A65: 
#> 
#>          Providing support to organizations whose primary activity is the
#>          production of plays. (Organizations that present the productions of
#>          others should be classified as presenters. (A61)
#> 
#>          A68: 
#> 
#>          Providing support to organizations whose primary activity is the
#>          production of musical events, including concert series and music
#>          festivals.
#> 
#>          A69: 
#> 
#>          Providing support to organizations whose primary activity is the
#>          operation of a symphony
#> 
#>          A6A: 
#> 
#>          
#> 
#>          A6B: 
#> 
#>          
#> 
#>          A6C: 
#> 
#>          
#> 
#>          A70: 
#> 
#>          Providing support to organizations that promote the study, teaching, or
#>          appreciation of the humanities (archaeology, art history, history,
#>          modern and classical languages, philosophy and ethics, and theology and
#>          comparative religion).
#> 
#>          A80: 
#> 
#>          Providing support to organizations that promote awareness of and
#>          appreciation for history and historical artifacts.  Use this code for
#>          organizations that are not specified below.
#> 
#>          A82: 
#> 
#>          Providing support to organizations that interpret, collect and preserve
#>          the historical heritage of a specific geographic location.
#> 
#>          A84: 
#> 
#>          
#> 
#>          A90: 
#> 
#>          Providing support to organizations that provide services such as
#>          facilities or touring assistance for arts organizations.
#> 
#>          A99: 
#> 
#>          
#> 
#>      RP: Research Institutes and/or Public Policy Analysis
#> 
#>          A00: 
#> 
#>          
#> 
#>          A34: 
#> 
#>          Providing support to organizations engaged in radio productions and
#>          broadcast, including public radio stations and radio stations that
#>          offer programming that is designed for community residents who are
#>          blind or visually impaired (radio reading services).
#> 
#>          A61: 
#> 
#>          Providing support to organizations that operate facilities including
#>          theaters for the performing arts.
#> 
#>  End of Preview